Refactor MainWindow UI and theme logic into mixins
