+++ codes/mainwindow.py
@@ -45,6 +45,13 @@ from workers.DEWorker import DEWorker
 from workers.SAWorker import SAWorker
 from workers.CMAESWorker import CMAESWorker
 # RL module import removed
+from gui.widgets import ModernQTabWidget, SidebarButton
+from gui.menu_mixin import MenuMixin
+from gui.beam_mixin import ContinuousBeamMixin
+from gui.microchip_mixin import MicrochipPageMixin
+from gui.sidebar_mixin import SidebarMixin
+from gui.theme_mixin import ThemeMixin
+from gui.stochastic_design_mixin import StochasticDesignMixin
 
 
 # Additional libraries used
@@ -73,46 +80,9 @@ except ImportError:
     print("Warning: Continuous Beam module imports failed. This functionality will be disabled.")
 
 
-class ModernQTabWidget(QTabWidget):
-    """Custom TabWidget with modern styling"""
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setDocumentMode(True)
-        self.setTabPosition(QTabWidget.North)
-        self.setMovable(True)
-
-
-class SidebarButton(QWidget):
-    """Custom styled sidebar button"""
-    def __init__(self, icon_path, text, parent=None):
-        super().__init__(parent)
-        layout = QHBoxLayout(self)
-        layout.setContentsMargins(10, 5, 10, 5)
-        
-        if icon_path:
-            icon = QLabel()
-            pixmap = QPixmap(icon_path)
-            icon.setPixmap(pixmap.scaled(24, 24, Qt.KeepAspectRatio, Qt.SmoothTransformation))
-            layout.addWidget(icon)
-        
-        label = QLabel(text)
-        label.setFont(QFont("Segoe UI", 10))
-        layout.addWidget(label)
-        layout.addStretch()
-        
-        self.setMinimumHeight(50)
-        self.setCursor(QCursor(Qt.PointingHandCursor))
-    
-    def enterEvent(self, event):
-        self.setStyleSheet("background-color: rgba(255, 255, 255, 0.2); border-radius: 5px;")
-        super().enterEvent(event)
-    
-    def leaveEvent(self, event):
-        self.setStyleSheet("")
-        super().leaveEvent(event)
         
 
-class MainWindow(QMainWindow):
+class MainWindow(QMainWindow, MenuMixin, ContinuousBeamMixin, MicrochipPageMixin, SidebarMixin, ThemeMixin, StochasticDesignMixin):
     def __init__(self):
         super().__init__()
         self.setWindowTitle("DeVana")
@@ -178,2606 +148,6 @@ class MainWindow(QMainWindow):
         """Print info about an array for debugging"""
         print(f"DEBUG {name}: type={type(array)}, shape={getattr(array, 'shape', 'N/A')}, min={np.min(array) if hasattr(array, '__len__') else array}, max={np.max(array) if hasattr(array, '__len__') else array}")
     
-    def create_sidebar(self):
-        """Create the sidebar with navigation buttons"""
-        # Create sidebar container
-        sidebar_container = QWidget()
-        sidebar_container.setObjectName("sidebar")
-        sidebar_container.setFixedWidth(250)
-        
-        sidebar_layout = QVBoxLayout(sidebar_container)
-        sidebar_layout.setContentsMargins(0, 0, 0, 0)
-        sidebar_layout.setSpacing(0)
-        
-        # Add logo/title at the top
-        logo_container = QWidget()
-        logo_container.setObjectName("logo-container")
-        logo_container.setMinimumHeight(100)
-        
-        logo_layout = QVBoxLayout(logo_container)
-        title = QLabel("DeVana")
-        title.setAlignment(Qt.AlignCenter)
-        title.setFont(QFont("Segoe UI", 18, QFont.Bold))
-        logo_layout.addWidget(title)
-        
-        version = QLabel("v2.0")
-        version.setAlignment(Qt.AlignCenter)
-        version.setFont(QFont("Segoe UI", 10))
-        logo_layout.addWidget(version)
-        
-        sidebar_layout.addWidget(logo_container)
-        
-        # Add separator
-        line = QFrame()
-        line.setFrameShape(QFrame.HLine)
-        line.setFrameShadow(QFrame.Sunken)
-        sidebar_layout.addWidget(line)
-        
-        # Main navigation buttons
-        nav_container = QWidget()
-        nav_layout = QVBoxLayout(nav_container)
-        nav_layout.setContentsMargins(10, 20, 10, 20)
-        nav_layout.setSpacing(10)
-        
-        # Stochastic Design button
-        self.stochastic_btn = SidebarButton(None, "Stochastic Design")
-        self.stochastic_btn.setObjectName("active-nav-btn")
-        self.stochastic_btn.mousePressEvent = lambda event: self.change_page(0)
-        nav_layout.addWidget(self.stochastic_btn)
-        
-        # Microchip Controller button
-        self.microchip_btn = SidebarButton(None, "Microchip Controller")
-        self.microchip_btn.mousePressEvent = lambda event: self.change_page(1)
-        nav_layout.addWidget(self.microchip_btn)
-        
-        # Continuous Beam button
-        self.beam_btn = SidebarButton(None, "Continuous Beam")
-        self.beam_btn.mousePressEvent = lambda event: self.change_page(2)
-        if not BEAM_IMPORTS_SUCCESSFUL:
-            self.beam_btn.setEnabled(False)
-            self.beam_btn.setToolTip("Continuous Beam module not available")
-        nav_layout.addWidget(self.beam_btn)
-        
-        nav_layout.addStretch()
-        sidebar_layout.addWidget(nav_container)
-        
-        # Theme toggle and settings at bottom
-        bottom_container = QWidget()
-        bottom_layout = QHBoxLayout(bottom_container)
-        
-        # Theme toggle button
-        self.theme_toggle = QPushButton("Toggle Theme")
-        self.theme_toggle.clicked.connect(self.toggle_theme)
-        bottom_layout.addWidget(self.theme_toggle)
-        
-        sidebar_layout.addWidget(bottom_container)
-        sidebar_layout.addSpacing(20)
-        
-        # Add sidebar to main layout
-        self.main_layout.addWidget(sidebar_container)
-    
-    def change_page(self, index):
-        """Change the active page in the content stack"""
-        self.content_stack.setCurrentIndex(index)
-        
-        # Update active button styling
-        for btn in [self.stochastic_btn, self.microchip_btn, self.beam_btn]:
-            btn.setObjectName("")
-            btn.setStyleSheet("")
-        
-        if index == 0:
-            self.stochastic_btn.setObjectName("active-nav-btn")
-        elif index == 1:
-            self.microchip_btn.setObjectName("active-nav-btn")
-        elif index == 2:
-            self.beam_btn.setObjectName("active-nav-btn")
-        
-        self.apply_current_theme()  # Reapply theme to update active button styling
-    
-    def toggle_theme(self):
-        """Toggle between light and dark themes"""
-        if self.current_theme == 'Dark':
-            self.current_theme = 'Light'
-            self.apply_light_theme()
-        else:
-            self.current_theme = 'Dark'
-            self.apply_dark_theme()
-    
-    def apply_current_theme(self):
-        """Apply the current theme"""
-        if self.current_theme == 'Dark':
-            self.apply_dark_theme()
-        else:
-            self.apply_light_theme()
-
-    def apply_dark_theme(self):
-        """Apply a modern dark theme with accent colors"""
-        dark_palette = QPalette()
-        
-        # Base colors
-        dark_color = QColor(18, 18, 18)
-        darker_color = QColor(12, 12, 12)
-        medium_color = QColor(40, 40, 40)
-        light_color = QColor(60, 60, 60)
-        text_color = QColor(240, 240, 240)
-        disabled_text_color = QColor(128, 128, 128)
-        
-        # Accent colors
-        primary_color = QColor(75, 111, 247)      # Blue
-        secondary_color = QColor(107, 64, 216)    # Purple
-        success_color = QColor(46, 204, 113)      # Green
-        warning_color = QColor(241, 196, 15)      # Yellow
-        danger_color = QColor(231, 76, 60)        # Red
-        info_color = QColor(52, 152, 219)         # Light Blue
-        
-        # Set up the palette
-        dark_palette.setColor(QPalette.Window, dark_color)
-        dark_palette.setColor(QPalette.WindowText, text_color)
-        dark_palette.setColor(QPalette.Base, darker_color)
-        dark_palette.setColor(QPalette.AlternateBase, medium_color)
-        dark_palette.setColor(QPalette.ToolTipBase, primary_color)
-        dark_palette.setColor(QPalette.ToolTipText, text_color)
-        dark_palette.setColor(QPalette.Text, text_color)
-        dark_palette.setColor(QPalette.Disabled, QPalette.Text, disabled_text_color)
-        dark_palette.setColor(QPalette.Button, dark_color)
-        dark_palette.setColor(QPalette.ButtonText, text_color)
-        dark_palette.setColor(QPalette.Disabled, QPalette.ButtonText, disabled_text_color)
-        dark_palette.setColor(QPalette.BrightText, Qt.red)
-        dark_palette.setColor(QPalette.Link, primary_color)
-        dark_palette.setColor(QPalette.Highlight, primary_color)
-        dark_palette.setColor(QPalette.HighlightedText, Qt.white)
-        dark_palette.setColor(QPalette.Disabled, QPalette.Highlight, QColor(80, 80, 80))
-        dark_palette.setColor(QPalette.Disabled, QPalette.HighlightedText, disabled_text_color)
-        
-        self.setPalette(dark_palette)
-        
-        # Dark theme stylesheet
-        dark_stylesheet = f"""
-            QMainWindow {{
-                background-color: #121212;
-            }}
-            QWidget {{
-                color: #F0F0F0;
-                background-color: #121212;
-            }}
-            #sidebar {{
-                background-color: #1A1A1A;
-                border-right: 1px solid #2D2D2D;
-            }}
-            #logo-container {{
-                background-color: #171717;
-                border-bottom: 1px solid #2D2D2D;
-            }}
-            #run-card {{
-                background-color: #1E1E1E;
-                border-radius: 8px;
-                border: 1px solid #2D2D2D;
-            }}
-            QTabWidget::pane {{
-                border: 1px solid #2D2D2D;
-                border-radius: 4px;
-                background-color: #1A1A1A;
-                top: -1px;
-            }}
-            QTabBar::tab {{
-                background-color: #1D1D1D;
-                color: #B0B0B0;
-                border: 1px solid #2D2D2D;
-                border-bottom: none;
-                border-top-left-radius: 4px;
-                border-top-right-radius: 4px;
-                padding: 8px 16px;
-                margin-right: 2px;
-            }}
-            QTabBar::tab:selected {{
-                background-color: #1A1A1A;
-                color: #FFFFFF;
-                border-bottom: 2px solid {primary_color.name()};
-            }}
-            QTabBar::tab:hover {{
-                background-color: #2D2D2D;
-            }}
-            QScrollArea, QScrollBar {{
-                border: none;
-                background-color: #1A1A1A;
-            }}
-            QScrollBar:vertical {{
-                border: none;
-                background-color: #1A1A1A;
-                width: 10px;
-                margin: 0px;
-            }}
-            QScrollBar:horizontal {{
-                border: none;
-                background-color: #1A1A1A;
-                height: 10px;
-                margin: 0px;
-            }}
-            QScrollBar::handle:vertical, QScrollBar::handle:horizontal {{
-                background-color: #3D3D3D;
-                border-radius: 5px;
-            }}
-            QScrollBar::handle:vertical:hover, QScrollBar::handle:horizontal:hover {{
-                background-color: #4D4D4D;
-            }}
-            QScrollBar::add-line, QScrollBar::sub-line {{
-                background: none;
-                border: none;
-            }}
-            QScrollBar::up-arrow, QScrollBar::down-arrow, QScrollBar::left-arrow, QScrollBar::right-arrow {{
-                background: none;
-                border: none;
-                color: none;
-            }}
-            QScrollBar::add-page, QScrollBar::sub-page {{
-                background: none;
-            }}
-            QGroupBox {{
-                border: 1px solid #2D2D2D;
-                border-radius: 6px;
-                margin-top: 12px;
-                padding-top: 12px;
-                font-weight: bold;
-            }}
-            QGroupBox::title {{
-                subcontrol-origin: margin;
-                subcontrol-position: top left;
-                left: 10px;
-                padding: 0 5px;
-                color: #CCCCCC;
-            }}
-            QGroupBox:hover {{
-                border: 1px solid {primary_color.name()};
-            }}
-            QPushButton {{
-                background-color: #333333;
-                border: 1px solid #444444;
-                border-radius: 4px;
-                padding: 6px 16px;
-                color: #FFFFFF;
-                font-weight: bold;
-            }}
-            QPushButton:hover {{
-                background-color: #444444;
-                border: 1px solid #555555;
-            }}
-            QPushButton:pressed {{
-                background-color: #222222;
-            }}
-            QPushButton:disabled {{
-                background-color: #2A2A2A;
-                color: #666666;
-                border: 1px solid #393939;
-            }}
-            #primary-button {{
-                background-color: {primary_color.name()};
-                border: none;
-                color: white;
-                border-radius: 4px;
-                padding: 8px 20px;
-                font-weight: bold;
-            }}
-            #primary-button:hover {{
-                background-color: {QColor(primary_color.red() + 20, primary_color.green() + 20, primary_color.blue() + 20).name()};
-            }}
-            #primary-button:pressed {{
-                background-color: {QColor(primary_color.red() - 20, primary_color.green() - 20, primary_color.blue() - 20).name()};
-            }}
-            #secondary-button {{
-                background-color: transparent;
-                border: 1px solid {primary_color.name()};
-                color: {primary_color.name()};
-                border-radius: 4px;
-                padding: 8px 16px;
-            }}
-            #secondary-button:hover {{
-                background-color: rgba(75, 111, 247, 0.1);
-            }}
-            #secondary-button:pressed {{
-                background-color: rgba(75, 111, 247, 0.2);
-            }}
-            QLineEdit, QTextEdit, QPlainTextEdit, QSpinBox, QDoubleSpinBox, QComboBox {{
-                border: 1px solid #333333;
-                border-radius: 4px;
-                padding: 5px 8px;
-                background-color: #1E1E1E;
-                color: #FFFFFF;
-                selection-background-color: {primary_color.name()};
-            }}
-            QLineEdit:hover, QTextEdit:hover, QPlainTextEdit:hover, QSpinBox:hover, QDoubleSpinBox:hover, QComboBox:hover {{
-                border: 1px solid {primary_color.name()};
-            }}
-            QLineEdit:focus, QTextEdit:focus, QPlainTextEdit:focus, QSpinBox:focus, QDoubleSpinBox:focus, QComboBox:focus {{
-                border: 1px solid {primary_color.name()};
-                background-color: #252525;
-            }}
-            QLineEdit:disabled, QTextEdit:disabled, QPlainTextEdit:disabled, QSpinBox:disabled, QDoubleSpinBox:disabled, QComboBox:disabled {{
-                border: 1px solid #2A2A2A;
-                background-color: #1A1A1A;
-                color: #666666;
-            }}
-            QComboBox::drop-down {{
-                subcontrol-origin: padding;
-                subcontrol-position: top right;
-                width: 20px;
-                border-left: 1px solid #333333;
-                border-top-right-radius: 3px;
-                border-bottom-right-radius: 3px;
-            }}
-            QComboBox::down-arrow {{
-                width: 12px;
-                height: 12px;
-            }}
-            QComboBox QAbstractItemView {{
-                border: 1px solid #333333;
-                background-color: #1E1E1E;
-                selection-background-color: {primary_color.name()};
-            }}
-            QToolBar {{
-                background-color: #1A1A1A;
-                border-bottom: 1px solid #2D2D2D;
-                spacing: 2px;
-            }}
-            QToolButton {{
-                background-color: transparent;
-                border: none;
-                border-radius: 4px;
-                padding: 3px;
-                margin: 1px;
-            }}
-            QToolButton:hover {{
-                background-color: rgba(255, 255, 255, 0.1);
-            }}
-            QToolButton:pressed {{
-                background-color: rgba(255, 255, 255, 0.2);
-            }}
-            QStatusBar {{
-                background-color: #1A1A1A;
-                color: #B0B0B0;
-                border-top: 1px solid #2D2D2D;
-            }}
-            QMenuBar {{
-                background-color: #1A1A1A;
-                color: #F0F0F0;
-                border-bottom: 1px solid #2D2D2D;
-            }}
-            QMenuBar::item {{
-                background-color: transparent;
-                padding: 6px 10px;
-            }}
-            QMenuBar::item:selected {{
-                background-color: #333333;
-            }}
-            QMenu {{
-                background-color: #1E1E1E;
-                border: 1px solid #333333;
-            }}
-            QMenu::item {{
-                padding: 6px 20px 6px 20px;
-            }}
-            QMenu::item:selected {{
-                background-color: {primary_color.name()};
-            }}
-            QMenu::separator {{
-                height: 1px;
-                background-color: #333333;
-                margin: 5px 0px 5px 0px;
-            }}
-            QCheckBox {{
-                spacing: 5px;
-            }}
-            QCheckBox::indicator {{
-                width: 16px;
-                height: 16px;
-                border: 1px solid #444444;
-                border-radius: 3px;
-                background-color: #1E1E1E;
-            }}
-            QCheckBox::indicator:checked {{
-                background-color: {primary_color.name()};
-                border: 1px solid {primary_color.name()};
-            }}
-            QCheckBox::indicator:hover {{
-                border: 1px solid {primary_color.name()};
-            }}
-            QTableView {{
-                background-color: #1E1E1E;
-                alternate-background-color: #252525;
-                border: 1px solid #333333;
-                border-radius: 4px;
-                gridline-color: #333333;
-                selection-background-color: {primary_color.name()};
-                selection-color: #FFFFFF;
-            }}
-            QHeaderView::section {{
-                background-color: #252525;
-                color: #CCCCCC;
-                padding: 5px;
-                border: 1px solid #333333;
-            }}
-            QSplitter::handle {{
-                background-color: #2D2D2D;
-            }}
-            QSplitter::handle:hover {{
-                background-color: {primary_color.name()};
-            }}
-            QDockWidget {{
-                titlebar-close-icon: url(close.png);
-                titlebar-normal-icon: url(undock.png);
-            }}
-            QDockWidget::title {{
-                text-align: center;
-                background-color: #1A1A1A;
-                padding: 5px;
-                border-bottom: 1px solid #333333;
-            }}
-            QProgressBar {{
-                border: 1px solid #333333;
-                border-radius: 3px;
-                background-color: #1E1E1E;
-                text-align: center;
-                color: #FFFFFF;
-            }}
-            QProgressBar::chunk {{
-                background-color: {primary_color.name()};
-                width: 1px;
-            }}
-            QSlider::groove:horizontal {{
-                border: none;
-                height: 5px;
-                background-color: #333333;
-                border-radius: 2px;
-            }}
-            QSlider::handle:horizontal {{
-                background-color: {primary_color.name()};
-                border: none;
-                width: 14px;
-                height: 14px;
-                margin: -5px 0px;
-                border-radius: 7px;
-            }}
-            QSlider::handle:horizontal:hover {{
-                background-color: {QColor(primary_color.red() + 20, primary_color.green() + 20, primary_color.blue() + 20).name()};
-            }}
-            QToolTip {{
-                border: 1px solid #333333;
-                background-color: #1E1E1E;
-                color: #FFFFFF;
-                padding: 5px;
-                opacity: 220;
-            }}
-            #active-nav-btn {{
-                background-color: rgba(75, 111, 247, 0.2);
-                border-left: 3px solid {primary_color.name()};
-                border-radius: 5px;
-            }}
-        """
-        self.setStyleSheet(dark_stylesheet)
-
-    def apply_light_theme(self):
-        """Apply a modern light theme with accent colors"""
-        light_palette = QPalette()
-        
-        # Base colors
-        light_color = QColor(250, 250, 250)
-        lighter_color = QColor(255, 255, 255)
-        medium_color = QColor(240, 240, 240)
-        dark_color = QColor(225, 225, 225)
-        text_color = QColor(40, 40, 40)
-        disabled_text_color = QColor(150, 150, 150)
-        
-        # Accent colors
-        primary_color = QColor(66, 133, 244)      # Blue
-        secondary_color = QColor(103, 58, 183)    # Purple
-        success_color = QColor(76, 175, 80)       # Green
-        warning_color = QColor(255, 152, 0)       # Orange
-        danger_color = QColor(244, 67, 54)        # Red
-        info_color = QColor(33, 150, 243)         # Light Blue
-        
-        # Set up the palette
-        light_palette.setColor(QPalette.Window, light_color)
-        light_palette.setColor(QPalette.WindowText, text_color)
-        light_palette.setColor(QPalette.Base, lighter_color)
-        light_palette.setColor(QPalette.AlternateBase, medium_color)
-        light_palette.setColor(QPalette.ToolTipBase, primary_color)
-        light_palette.setColor(QPalette.ToolTipText, Qt.white)
-        light_palette.setColor(QPalette.Text, text_color)
-        light_palette.setColor(QPalette.Disabled, QPalette.Text, disabled_text_color)
-        light_palette.setColor(QPalette.Button, light_color)
-        light_palette.setColor(QPalette.ButtonText, text_color)
-        light_palette.setColor(QPalette.Disabled, QPalette.ButtonText, disabled_text_color)
-        light_palette.setColor(QPalette.BrightText, Qt.red)
-        light_palette.setColor(QPalette.Link, primary_color)
-        light_palette.setColor(QPalette.Highlight, primary_color)
-        light_palette.setColor(QPalette.HighlightedText, Qt.white)
-        light_palette.setColor(QPalette.Disabled, QPalette.Highlight, QColor(200, 200, 200))
-        light_palette.setColor(QPalette.Disabled, QPalette.HighlightedText, disabled_text_color)
-        
-        self.setPalette(light_palette)
-        
-        # Light theme stylesheet
-        light_stylesheet = f"""
-            QMainWindow {{
-                background-color: #FAFAFA;
-            }}
-            QWidget {{
-                color: #282828;
-                background-color: #FAFAFA;
-            }}
-            #sidebar {{
-                background-color: #F0F0F0;
-                border-right: 1px solid #E0E0E0;
-            }}
-            #logo-container {{
-                background-color: #FFFFFF;
-                border-bottom: 1px solid #E0E0E0;
-            }}
-            #run-card {{
-                background-color: #FFFFFF;
-                border-radius: 8px;
-                border: 1px solid #E0E0E0;
-                box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.05);
-            }}
-            QTabWidget::pane {{
-                border: 1px solid #E0E0E0;
-                border-radius: 4px;
-                background-color: #FFFFFF;
-                top: -1px;
-            }}
-            QTabBar::tab {{
-                background-color: #F5F5F5;
-                color: #707070;
-                border: 1px solid #E0E0E0;
-                border-bottom: none;
-                border-top-left-radius: 4px;
-                border-top-right-radius: 4px;
-                padding: 8px 16px;
-                margin-right: 2px;
-            }}
-            QTabBar::tab:selected {{
-                background-color: #FFFFFF;
-                color: #282828;
-                border-bottom: 2px solid {primary_color.name()};
-            }}
-            QTabBar::tab:hover {{
-                background-color: #EEEEEE;
-            }}
-            QScrollArea, QScrollBar {{
-                border: none;
-                background-color: #FFFFFF;
-            }}
-            QScrollBar:vertical {{
-                border: none;
-                background-color: #FAFAFA;
-                width: 10px;
-                margin: 0px;
-            }}
-            QScrollBar:horizontal {{
-                border: none;
-                background-color: #FAFAFA;
-                height: 10px;
-                margin: 0px;
-            }}
-            QScrollBar::handle:vertical, QScrollBar::handle:horizontal {{
-                background-color: #D0D0D0;
-                border-radius: 5px;
-            }}
-            QScrollBar::handle:vertical:hover, QScrollBar::handle:horizontal:hover {{
-                background-color: #B0B0B0;
-            }}
-            QScrollBar::add-line, QScrollBar::sub-line {{
-                background: none;
-                border: none;
-            }}
-            QScrollBar::up-arrow, QScrollBar::down-arrow, QScrollBar::left-arrow, QScrollBar::right-arrow {{
-                background: none;
-                border: none;
-                color: none;
-            }}
-            QScrollBar::add-page, QScrollBar::sub-page {{
-                background: none;
-            }}
-            QGroupBox {{
-                border: 1px solid #E0E0E0;
-                border-radius: 6px;
-                margin-top: 12px;
-                padding-top: 12px;
-                font-weight: bold;
-                background-color: #FFFFFF;
-            }}
-            QGroupBox::title {{
-                subcontrol-origin: margin;
-                subcontrol-position: top left;
-                left: 10px;
-                padding: 0 5px;
-                color: #505050;
-            }}
-            QGroupBox:hover {{
-                border: 1px solid {primary_color.name()};
-            }}
-            QPushButton {{
-                background-color: #F0F0F0;
-                border: 1px solid #D0D0D0;
-                border-radius: 4px;
-                padding: 6px 16px;
-                color: #404040;
-                font-weight: bold;
-            }}
-            QPushButton:hover {{
-                background-color: #E5E5E5;
-                border: 1px solid #C0C0C0;
-            }}
-            QPushButton:pressed {{
-                background-color: #DEDEDE;
-            }}
-            QPushButton:disabled {{
-                background-color: #F5F5F5;
-                color: #B0B0B0;
-                border: 1px solid #E0E0E0;
-            }}
-            #primary-button {{
-                background-color: {primary_color.name()};
-                border: none;
-                color: white;
-                border-radius: 4px;
-                padding: 8px 20px;
-                font-weight: bold;
-            }}
-            #primary-button:hover {{
-                background-color: {QColor(primary_color.red() + 20, primary_color.green() + 20, primary_color.blue() + 20).name()};
-            }}
-            #primary-button:pressed {{
-                background-color: {QColor(primary_color.red() - 20, primary_color.green() - 20, primary_color.blue() - 20).name()};
-            }}
-            #secondary-button {{
-                background-color: transparent;
-                border: 1px solid {primary_color.name()};
-                color: {primary_color.name()};
-                border-radius: 4px;
-                padding: 8px 16px;
-            }}
-            #secondary-button:hover {{
-                background-color: rgba(66, 133, 244, 0.1);
-            }}
-            #secondary-button:pressed {{
-                background-color: rgba(66, 133, 244, 0.2);
-            }}
-            QLineEdit, QTextEdit, QPlainTextEdit, QSpinBox, QDoubleSpinBox, QComboBox {{
-                border: 1px solid #E0E0E0;
-                border-radius: 4px;
-                padding: 5px 8px;
-                background-color: #FFFFFF;
-                color: #202020;
-                selection-background-color: {primary_color.name()};
-            }}
-            QLineEdit:hover, QTextEdit:hover, QPlainTextEdit:hover, QSpinBox:hover, QDoubleSpinBox:hover, QComboBox:hover {{
-                border: 1px solid {primary_color.name()};
-            }}
-            QLineEdit:focus, QTextEdit:focus, QPlainTextEdit:focus, QSpinBox:focus, QDoubleSpinBox:focus, QComboBox:focus {{
-                border: 1px solid {primary_color.name()};
-                background-color: #FFFFFF;
-            }}
-            QLineEdit:disabled, QTextEdit:disabled, QPlainTextEdit:disabled, QSpinBox:disabled, QDoubleSpinBox:disabled, QComboBox:disabled {{
-                border: 1px solid #E5E5E5;
-                background-color: #F5F5F5;
-                color: #B0B0B0;
-            }}
-            QComboBox::drop-down {{
-                subcontrol-origin: padding;
-                subcontrol-position: top right;
-                width: 20px;
-                border-left: 1px solid #E0E0E0;
-                border-top-right-radius: 3px;
-                border-bottom-right-radius: 3px;
-            }}
-            QComboBox::down-arrow {{
-                width: 12px;
-                height: 12px;
-            }}
-            QComboBox QAbstractItemView {{
-                border: 1px solid #E0E0E0;
-                background-color: #FFFFFF;
-                selection-background-color: {primary_color.name()};
-            }}
-            QToolBar {{
-                background-color: #F5F5F5;
-                border-bottom: 1px solid #E0E0E0;
-                spacing: 2px;
-            }}
-            QToolButton {{
-                background-color: transparent;
-                border: none;
-                border-radius: 4px;
-                padding: 3px;
-                margin: 1px;
-            }}
-            QToolButton:hover {{
-                background-color: rgba(0, 0, 0, 0.05);
-            }}
-            QToolButton:pressed {{
-                background-color: rgba(0, 0, 0, 0.1);
-            }}
-            QStatusBar {{
-                background-color: #F5F5F5;
-                color: #606060;
-                border-top: 1px solid #E0E0E0;
-            }}
-            QMenuBar {{
-                background-color: #F5F5F5;
-                color: #404040;
-                border-bottom: 1px solid #E0E0E0;
-            }}
-            QMenuBar::item {{
-                background-color: transparent;
-                padding: 6px 10px;
-            }}
-            QMenuBar::item:selected {{
-                background-color: #E5E5E5;
-            }}
-            QMenu {{
-                background-color: #FFFFFF;
-                border: 1px solid #E0E0E0;
-            }}
-            QMenu::item {{
-                padding: 6px 20px 6px 20px;
-            }}
-            QMenu::item:selected {{
-                background-color: {primary_color.name()};
-                color: white;
-            }}
-            QMenu::separator {{
-                height: 1px;
-                background-color: #E0E0E0;
-                margin: 5px 0px 5px 0px;
-            }}
-            QCheckBox {{
-                spacing: 5px;
-            }}
-            QCheckBox::indicator {{
-                width: 16px;
-                height: 16px;
-                border: 1px solid #D0D0D0;
-                border-radius: 3px;
-                background-color: #FFFFFF;
-            }}
-            QCheckBox::indicator:checked {{
-                background-color: {primary_color.name()};
-                border: 1px solid {primary_color.name()};
-            }}
-            QCheckBox::indicator:hover {{
-                border: 1px solid {primary_color.name()};
-            }}
-            QTableView {{
-                background-color: #FFFFFF;
-                alternate-background-color: #F9F9F9;
-                border: 1px solid #E0E0E0;
-                border-radius: 4px;
-                gridline-color: #E0E0E0;
-                selection-background-color: {primary_color.name()};
-                selection-color: white;
-            }}
-            QHeaderView::section {{
-                background-color: #F0F0F0;
-                color: #404040;
-                padding: 5px;
-                border: 1px solid #E0E0E0;
-            }}
-            QSplitter::handle {{
-                background-color: #E0E0E0;
-            }}
-            QSplitter::handle:hover {{
-                background-color: {primary_color.name()};
-            }}
-            QDockWidget {{
-                titlebar-close-icon: url(close.png);
-                titlebar-normal-icon: url(undock.png);
-            }}
-            QDockWidget::title {{
-                text-align: center;
-                background-color: #F5F5F5;
-                padding: 5px;
-                border-bottom: 1px solid #E0E0E0;
-            }}
-            QProgressBar {{
-                border: 1px solid #E0E0E0;
-                border-radius: 3px;
-                background-color: #FFFFFF;
-                text-align: center;
-                color: #404040;
-            }}
-            QProgressBar::chunk {{
-                background-color: {primary_color.name()};
-                width: 1px;
-            }}
-            QSlider::groove:horizontal {{
-                border: none;
-                height: 5px;
-                background-color: #E0E0E0;
-                border-radius: 2px;
-            }}
-            QSlider::handle:horizontal {{
-                background-color: {primary_color.name()};
-                border: none;
-                width: 14px;
-                height: 14px;
-                margin: -5px 0px;
-                border-radius: 7px;
-            }}
-            QSlider::handle:horizontal:hover {{
-                background-color: {QColor(primary_color.red() + 20, primary_color.green() + 20, primary_color.blue() + 20).name()};
-            }}
-            QToolTip {{
-                border: 1px solid #E0E0E0;
-                background-color: #FFFFFF;
-                color: #404040;
-                padding: 5px;
-                opacity: 220;
-            }}
-            #active-nav-btn {{
-                background-color: rgba(66, 133, 244, 0.1);
-                border-left: 3px solid {primary_color.name()};
-                border-radius: 5px;
-            }}
-        """
-        self.setStyleSheet(light_stylesheet)
-
-    def create_stochastic_design_page(self):
-        """Create the stochastic design page with all existing tabs"""
-        stochastic_page = QWidget()
-        page_layout = QVBoxLayout(stochastic_page)
-        page_layout.setContentsMargins(20, 20, 20, 20)
-        
-        # Title and description
-        header = QWidget()
-        header_layout = QVBoxLayout(header)
-        title = QLabel("Stochastic Design")
-        title.setFont(QFont("Segoe UI", 20, QFont.Bold))
-        header_layout.addWidget(title)
-        
-        description = QLabel("Design and optimize stochastic vibration systems with advanced algorithms")
-        description.setFont(QFont("Segoe UI", 11))
-        header_layout.addWidget(description)
-        
-        # Add header to page layout
-        page_layout.addWidget(header)
-        
-        # Create a horizontal split for content and results
-        content_splitter = QSplitter(Qt.Horizontal)
-        
-        # Left panel - Tabs and main controls
-        left_panel = QWidget()
-        left_layout = QVBoxLayout(left_panel)
-        left_layout.setContentsMargins(0, 0, 0, 0)
-        
-        # Create all original tabs
-        self.create_main_system_tab()
-        self.create_dva_parameters_tab()
-        self.create_target_weights_tab()
-        self.create_frequency_tab()
-        self.create_omega_sensitivity_tab()  # Create the new Omega Sensitivity tab
-        self.create_sobol_analysis_tab()
-        self.create_ga_tab()
-        self.create_pso_tab()
-        self.create_de_tab()
-        self.create_sa_tab()
-        self.create_cmaes_tab()
-        # RL tab creation removed
-        
-        # Create the tab widget with modern styling
-        self.design_tabs = ModernQTabWidget()
-        
-        # Create mother tab for Input (contains Main System, DVA Parameters, Targets & Weights, Frequency Plot)
-        self.input_tabs = ModernQTabWidget()
-        self.input_tabs.addTab(self.main_system_tab, "Main System")
-        self.input_tabs.addTab(self.dva_tab, "DVA Parameters")
-        self.input_tabs.addTab(self.tw_tab, "Targets & Weights")
-        self.input_tabs.addTab(self.freq_tab, "Frequency & Plot")
-        self.input_tabs.addTab(self.omega_sensitivity_tab, "Ω Sensitivity")
-
-        # Create mother tab for Sensitivity Analysis (contains Sobol Analysis)
-        self.sensitivity_tabs = ModernQTabWidget()
-        self.sensitivity_tabs.addTab(self.sobol_tab, "Sobol Analysis")
-
-        # Create mother tab for Optimization (contains GA, PSO, DE, SA, and CMA-ES)
-        self.optimization_tabs = ModernQTabWidget()
-        self.optimization_tabs.addTab(self.ga_tab, "GA Optimization")
-        self.optimization_tabs.addTab(self.pso_tab, "PSO Optimization")
-        self.optimization_tabs.addTab(self.de_tab, "DE Optimization")
-        self.optimization_tabs.addTab(self.sa_tab, "SA Optimization")
-        self.optimization_tabs.addTab(self.cmaes_tab, "CMA-ES Optimization")
-
-        # Comprehensive Analysis tab has been removed
-        # Here we add the RL tab (with integrated Sobol settings, epsilon decay, and reward settings sub-tabs)
-        # Comprehensive analysis tab creation removed
-        # RL tab addition removed
-        
-        # Add all tabs to the main tab widget
-        self.design_tabs.addTab(self.input_tabs, "Input")
-        self.design_tabs.addTab(self.sensitivity_tabs, "Sensitivity Analysis")
-        self.design_tabs.addTab(self.optimization_tabs, "Optimization")
-        # Comprehensive tab addition removed
-        
-        left_layout.addWidget(self.design_tabs)
-        
-        # Run buttons in a card-like container
-        run_card = QFrame()
-        run_card.setObjectName("run-card")
-        run_card.setMinimumHeight(120)
-        run_card_layout = QVBoxLayout(run_card)
-        
-        run_title = QLabel("Actions")
-        run_title.setFont(QFont("Segoe UI", 12, QFont.Bold))
-        run_card_layout.addWidget(run_title)
-        
-        run_buttons_layout = QHBoxLayout()
-        
-        self.run_frf_button = QPushButton("Run FRF")
-        self.run_frf_button.setObjectName("primary-button")
-        self.run_frf_button.setMinimumHeight(40)
-        self.run_frf_button.clicked.connect(self.run_frf)
-        self.run_frf_button.setVisible(False)  # Hide button
-        
-        self.run_sobol_button = QPushButton("Run Sobol")
-        self.run_sobol_button.setObjectName("primary-button")
-        self.run_sobol_button.setMinimumHeight(40)
-        self.run_sobol_button.clicked.connect(self.run_sobol)
-        self.run_sobol_button.setVisible(False)  # Hide button
-        
-        self.run_ga_button = QPushButton("Run GA")
-        self.run_ga_button.setObjectName("primary-button")
-        self.run_ga_button.setMinimumHeight(40)
-        self.run_ga_button.clicked.connect(self.run_ga)
-        self.run_ga_button.setVisible(False)  # Hide button
-        
-        self.run_pso_button = QPushButton("Run PSO")
-        self.run_pso_button.setObjectName("primary-button")
-        self.run_pso_button.setMinimumHeight(40)
-        self.run_pso_button.clicked.connect(self.run_pso)
-        self.run_pso_button.setVisible(False)  # Hide button
-        
-        self.run_de_button = QPushButton("Run DE")
-        self.run_de_button.setObjectName("primary-button")
-        self.run_de_button.setMinimumHeight(40)
-        self.run_de_button.clicked.connect(self.run_de)
-        self.run_de_button.setVisible(False)  # Hide button
-        
-        self.run_sa_button = QPushButton("Run SA")
-        self.run_sa_button.setObjectName("primary-button")
-        self.run_sa_button.setMinimumHeight(40)
-        self.run_sa_button.clicked.connect(self.run_sa)
-        self.run_sa_button.setVisible(False)  # Hide button
-        
-        self.run_cmaes_button = QPushButton("Run CMA-ES")
-        self.run_cmaes_button.setObjectName("primary-button")
-        self.run_cmaes_button.setMinimumHeight(40)
-        self.run_cmaes_button.clicked.connect(self.run_cmaes)
-        self.run_cmaes_button.setVisible(False)  # Hide button
-        
-        run_buttons_layout.addWidget(self.run_frf_button)
-        run_buttons_layout.addWidget(self.run_sobol_button)
-        run_buttons_layout.addWidget(self.run_ga_button)
-        run_buttons_layout.addWidget(self.run_pso_button)
-        run_buttons_layout.addWidget(self.run_de_button)
-        run_buttons_layout.addWidget(self.run_sa_button)
-        run_buttons_layout.addWidget(self.run_cmaes_button)
-        
-        run_card_layout.addLayout(run_buttons_layout)
-        run_card.setVisible(False)  # Hide entire run card
-        left_layout.addWidget(run_card)
-        
-        # Add left panel to splitter
-        content_splitter.addWidget(left_panel)
-        
-        # Right panel - Results area with tabs
-        right_panel = QWidget()
-        right_layout = QVBoxLayout(right_panel)
-        right_layout.setContentsMargins(0, 0, 0, 0)
-        
-        results_tabs = ModernQTabWidget()
-        
-        # Results text panel
-        results_panel = QWidget()
-        results_panel_layout = QVBoxLayout(results_panel)
-        
-        results_title = QLabel("Results")
-        results_title.setFont(QFont("Segoe UI", 12, QFont.Bold))
-        results_panel_layout.addWidget(results_title)
-        
-        self.results_text = QTextEdit()
-        self.results_text.setReadOnly(True)
-        
-        results_panel_layout.addWidget(self.results_text)
-        
-        # FRF plots panel
-        frf_panel = QWidget()
-        frf_layout = QVBoxLayout(frf_panel)
-        
-        frf_header = QWidget()
-        frf_header_layout = QHBoxLayout(frf_header)
-        
-        frf_title = QLabel("FRF Plots")
-        frf_title.setFont(QFont("Segoe UI", 12, QFont.Bold))
-        frf_header_layout.addWidget(frf_title)
-        
-        self.frf_combo = QComboBox()
-        self.frf_combo.currentIndexChanged.connect(self.update_frf_plot)
-        frf_header_layout.addWidget(self.frf_combo)
-        
-        self.frf_save_plot_button = QPushButton("Save Plot")
-        self.frf_save_plot_button.setObjectName("secondary-button")
-        self.frf_save_plot_button.clicked.connect(lambda: self.save_plot(self.frf_fig, "FRF"))
-        frf_header_layout.addWidget(self.frf_save_plot_button)
-        
-        frf_layout.addWidget(frf_header)
-        
-        self.frf_fig = Figure(figsize=(6, 4))
-        self.frf_canvas = FigureCanvas(self.frf_fig)
-        self.frf_canvas.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-        
-        self.frf_toolbar = NavigationToolbar(self.frf_canvas, frf_panel)
-        frf_layout.addWidget(self.frf_toolbar)
-        frf_layout.addWidget(self.frf_canvas)
-        
-        # Comparative FRF plots panel
-        comp_panel = QWidget()
-        comp_layout = QVBoxLayout(comp_panel)
-        
-        comp_header = QWidget()
-        comp_header_layout = QHBoxLayout(comp_header)
-        
-        comp_title = QLabel("Comparative FRF Plots")
-        comp_title.setFont(QFont("Segoe UI", 12, QFont.Bold))
-        comp_header_layout.addWidget(comp_title)
-        
-        self.comp_save_plot_button = QPushButton("Save Plot")
-        self.comp_save_plot_button.setObjectName("secondary-button")
-        self.comp_save_plot_button.clicked.connect(lambda: self.save_plot(self.comp_fig, "Comparative FRF"))
-        comp_header_layout.addWidget(self.comp_save_plot_button)
-        
-        comp_layout.addWidget(comp_header)
-        
-        self.comp_fig = Figure(figsize=(6, 4))
-        self.comp_canvas = FigureCanvas(self.comp_fig)
-        self.comp_canvas.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-        
-        self.comp_toolbar = NavigationToolbar(self.comp_canvas, comp_panel)
-        comp_layout.addWidget(self.comp_toolbar)
-        comp_layout.addWidget(self.comp_canvas)
-        
-        # Sobol plots panel
-        sobol_panel = QWidget()
-        sobol_layout = QVBoxLayout(sobol_panel)
-        
-        sobol_header = QWidget()
-        sobol_header_layout = QHBoxLayout(sobol_header)
-        
-        sobol_title = QLabel("Sobol Analysis")
-        sobol_title.setFont(QFont("Segoe UI", 12, QFont.Bold))
-        sobol_header_layout.addWidget(sobol_title)
-        
-        self.sobol_combo = QComboBox()
-        self.sobol_combo.currentIndexChanged.connect(self.update_sobol_plot)
-        sobol_header_layout.addWidget(self.sobol_combo)
-        
-        self.sobol_save_plot_button = QPushButton("Save Plot")
-        self.sobol_save_plot_button.setObjectName("secondary-button")
-        self.sobol_save_plot_button.clicked.connect(lambda: self.save_plot(self.sobol_fig, "Sobol"))
-        sobol_header_layout.addWidget(self.sobol_save_plot_button)
-        
-        sobol_layout.addWidget(sobol_header)
-        
-        self.sobol_fig = Figure(figsize=(6, 4))
-        self.sobol_canvas = FigureCanvas(self.sobol_fig)
-        self.sobol_canvas.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-        
-        self.sobol_toolbar = NavigationToolbar(self.sobol_canvas, sobol_panel)
-        sobol_layout.addWidget(self.sobol_toolbar)
-        sobol_layout.addWidget(self.sobol_canvas)
-        
-        sobol_results_container = QWidget()
-        sobol_results_layout = QVBoxLayout(sobol_results_container)
-        sobol_results_layout.setContentsMargins(0, 10, 0, 0)
-        
-        sobol_results_header = QHBoxLayout()
-        sobol_results_title = QLabel("Sobol Results")
-        sobol_results_title.setFont(QFont("Segoe UI", 10, QFont.Bold))
-        sobol_results_header.addWidget(sobol_results_title)
-        
-        self.sobol_save_results_button = QPushButton("Save Results")
-        self.sobol_save_results_button.setObjectName("secondary-button")
-        self.sobol_save_results_button.clicked.connect(self.save_sobol_results)
-        sobol_results_header.addWidget(self.sobol_save_results_button)
-        
-        sobol_results_layout.addLayout(sobol_results_header)
-        
-        self.sobol_results_text = QTextEdit()
-        self.sobol_results_text.setReadOnly(True)
-        self.sobol_results_text.setStyleSheet("font-family: monospace;")
-        sobol_results_layout.addWidget(self.sobol_results_text)
-        
-        sobol_layout.addWidget(sobol_results_container)
-        
-        # Add tabs to results panel
-        results_tabs.addTab(results_panel, "Text Results")
-        results_tabs.addTab(frf_panel, "FRF Visualization")
-        results_tabs.addTab(comp_panel, "Comparative FRF")
-        results_tabs.addTab(sobol_panel, "Sobol Visualization")
-        
-        right_layout.addWidget(results_tabs)
-        
-        # Add right panel to splitter
-        content_splitter.addWidget(right_panel)
-        
-        # Set the initial sizes of the splitter
-        content_splitter.setSizes([800, 800])
-        
-        # Add the splitter to the page layout
-        page_layout.addWidget(content_splitter)
-        
-        # Add page to content stack
-        self.content_stack.addWidget(stochastic_page)
-
-    def create_main_system_tab(self):
-        """Create the main system parameters tab"""
-        self.main_system_tab = QWidget()
-        layout = QVBoxLayout(self.main_system_tab)
-        
-        # Create a scroll area for potentially large content
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        
-        # Create main container widget
-        main_container = QWidget()
-        main_layout = QVBoxLayout(main_container)
-        
-        # Main system parameters group
-        params_group = QGroupBox("Main System Parameters")
-        params_layout = QFormLayout(params_group)
-        
-        # MU parameter
-        self.mu_box = QDoubleSpinBox()
-        self.mu_box.setRange(-1e6, 1e6)
-        self.mu_box.setDecimals(6)
-        self.mu_box.setValue(1.0)
-        params_layout.addRow("μ (MU):", self.mu_box)
-
-        # LANDA parameters (Lambda)
-        self.landa_boxes = []
-        for i in range(5):
-            box = QDoubleSpinBox()
-            box.setRange(-1e6, 1e6)
-            box.setDecimals(6)
-            if i < 2:
-                box.setValue(1.0)
-            else:
-                box.setValue(0.5)
-            self.landa_boxes.append(box)
-            params_layout.addRow(f"Λ_{i+1}:", box)
-
-        # NU parameters
-        self.nu_boxes = []
-        for i in range(5):
-            box = QDoubleSpinBox()
-            box.setRange(-1e6, 1e6)
-            box.setDecimals(6)
-            box.setValue(0.75)
-            self.nu_boxes.append(box)
-            params_layout.addRow(f"Ν_{i+1}:", box)
-
-        # A_LOW parameter
-        self.a_low_box = QDoubleSpinBox()
-        self.a_low_box.setRange(0, 1e10)
-        self.a_low_box.setDecimals(6)
-        self.a_low_box.setValue(0.05)
-        params_layout.addRow("A_LOW:", self.a_low_box)
-
-        # A_UPP parameter
-        self.a_up_box = QDoubleSpinBox()
-        self.a_up_box.setRange(0, 1e10)
-        self.a_up_box.setDecimals(6)
-        self.a_up_box.setValue(0.05)
-        params_layout.addRow("A_UPP:", self.a_up_box)
-
-        # F_1 parameter
-        self.f_1_box = QDoubleSpinBox()
-        self.f_1_box.setRange(0, 1e10)
-        self.f_1_box.setDecimals(6)
-        self.f_1_box.setValue(100.0)
-        params_layout.addRow("F_1:", self.f_1_box)
-
-        # F_2 parameter
-        self.f_2_box = QDoubleSpinBox()
-        self.f_2_box.setRange(0, 1e10)
-        self.f_2_box.setDecimals(6)
-        self.f_2_box.setValue(100.0)
-        params_layout.addRow("F_2:", self.f_2_box)
-
-        # OMEGA_DC parameter
-        self.omega_dc_box = QDoubleSpinBox()
-        self.omega_dc_box.setRange(0, 1e10)
-        self.omega_dc_box.setDecimals(6)
-        self.omega_dc_box.setValue(5000.0)
-        params_layout.addRow("Ω_DC:", self.omega_dc_box)
-
-        # ZETA_DC parameter
-        self.zeta_dc_box = QDoubleSpinBox()
-        self.zeta_dc_box.setRange(0, 1e10)
-        self.zeta_dc_box.setDecimals(6)
-        self.zeta_dc_box.setValue(0.01)
-        params_layout.addRow("ζ_DC:", self.zeta_dc_box)
-        
-        main_layout.addWidget(params_group)
-        main_layout.addStretch()
-        
-        # Set the container as the scroll area's widget
-        scroll_area.setWidget(main_container)
-        
-        # Add scroll area to the tab's layout
-        layout.addWidget(scroll_area)
-
-    def create_dva_parameters_tab(self):
-        """Create the DVA parameters tab"""
-        self.dva_tab = QWidget()
-        layout = QVBoxLayout(self.dva_tab)
-        
-        # Create a scroll area for potentially large content
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        
-        # Create main container widget
-        main_container = QWidget()
-        main_layout = QVBoxLayout(main_container)
-        
-        # Add button to apply optimized DVA parameters from GA
-        apply_optimized_container = QWidget()
-        apply_optimized_layout = QHBoxLayout(apply_optimized_container)
-        apply_optimized_layout.setContentsMargins(0, 0, 0, 10)
-        
-        apply_optimized_button = QPushButton("Apply Optimized DVA Parameters")
-        apply_optimized_button.setToolTip("Apply the best parameters from the last optimization run")
-        apply_optimized_button.clicked.connect(self.apply_optimized_dva_parameters)
-        apply_optimized_layout.addWidget(apply_optimized_button)
-        
-        apply_optimizer_combo = QComboBox()
-        apply_optimizer_combo.addItems(["Genetic Algorithm (GA)", "Particle Swarm (PSO)", 
-                                       "Differential Evolution (DE)", "Simulated Annealing (SA)", 
-                                       "CMA-ES", "Reinforcement Learning (RL)"])
-        apply_optimized_layout.addWidget(apply_optimizer_combo)
-        self.dva_optimizer_combo = apply_optimizer_combo
-        
-        main_layout.addWidget(apply_optimized_container)
-        
-        # BETA parameters group
-        beta_group = QGroupBox("β (beta) Parameters")
-        beta_form = QFormLayout(beta_group)
-        self.beta_boxes = []
-        for i in range(15):
-            b = QDoubleSpinBox()
-            b.setRange(-1e6, 1e6)
-            b.setDecimals(6)
-            self.beta_boxes.append(b)
-            beta_form.addRow(f"β_{i+1}:", b)
-        main_layout.addWidget(beta_group)
-
-        # LAMBDA parameters group
-        lambda_group = QGroupBox("λ (lambda) Parameters")
-        lambda_form = QFormLayout(lambda_group)
-        self.lambda_boxes = []
-        for i in range(15):
-            l = QDoubleSpinBox()
-            l.setRange(-1e6, 1e6)
-            l.setDecimals(6)
-            self.lambda_boxes.append(l)
-            lambda_form.addRow(f"λ_{i+1}:", l)
-        main_layout.addWidget(lambda_group)
-
-        # MU parameters group
-        mu_group = QGroupBox("μ (mu) Parameters")
-        mu_form = QFormLayout(mu_group)
-        self.mu_dva_boxes = []
-        for i in range(3):
-            m = QDoubleSpinBox()
-            m.setRange(-1e6, 1e6)
-            m.setDecimals(6)
-            self.mu_dva_boxes.append(m)
-            mu_form.addRow(f"μ_{i+1}:", m)
-        main_layout.addWidget(mu_group)
-
-        # NU parameters group
-        nu_group = QGroupBox("ν (nu) Parameters")
-        nu_form = QFormLayout(nu_group)
-        self.nu_dva_boxes = []
-        for i in range(15):
-            n = QDoubleSpinBox()
-            n.setRange(-1e6, 1e6)
-            n.setDecimals(6)
-            self.nu_dva_boxes.append(n)
-            nu_form.addRow(f"ν_{i+1}:", n)
-        main_layout.addWidget(nu_group)
-        
-        main_layout.addStretch()
-        
-        # Set the container as the scroll area's widget
-        scroll_area.setWidget(main_container)
-        
-        # Add scroll area to the tab's layout
-        layout.addWidget(scroll_area)
-    
-    def create_target_weights_tab(self):
-        """Create the targets and weights tab"""
-        self.tw_tab = QWidget()
-        layout = QVBoxLayout(self.tw_tab)
-        
-        # Create a scroll area for potentially large content
-        scroll_area = QScrollArea()
-        scroll_area.setWidgetResizable(True)
-        
-        # Create main container widget
-        main_container = QWidget()
-        main_layout = QVBoxLayout(main_container)
-        
-        # Initialize dictionaries to store all target and weight spinboxes
-        self.mass_target_spins = {}
-        self.mass_weight_spins = {}
-
-        # Create groups for each mass
-        for mass_num in range(1, 6):
-            mass_group = QGroupBox(f"Mass {mass_num} Targets & Weights")
-            mg_layout = QVBoxLayout(mass_group)
-
-            # Peak values group
-            peak_group = QGroupBox("Peak Values & Weights")
-            peak_form = QFormLayout(peak_group)
-            for peak_num in range(1, 5):
-                pv = QDoubleSpinBox()
-                pv.setRange(0, 1e6)
-                pv.setDecimals(6)
-                wv = QDoubleSpinBox()
-                wv.setRange(0, 1e3)
-                wv.setDecimals(6)
-                
-                peak_form.addRow(f"Peak Value {peak_num}:", pv)
-                peak_form.addRow(f"Weight Peak Value {peak_num}:", wv)
-                
-                self.mass_target_spins[f"peak_value_{peak_num}_m{mass_num}"] = pv
-                self.mass_weight_spins[f"peak_value_{peak_num}_m{mass_num}"] = wv
-            mg_layout.addWidget(peak_group)
-            
-            # Peak positions group (in a separate section)
-            peak_pos_group = QGroupBox("Peak Positions & Weights")
-            peak_pos_form = QFormLayout(peak_pos_group)
-            for peak_num in range(1, 6):  # Note: 1-5 (not 1-4)
-                pp = QDoubleSpinBox()
-                pp.setRange(0, 1e6)
-                pp.setDecimals(6)
-                wpp = QDoubleSpinBox()
-                wpp.setRange(0, 1e3)
-                wpp.setDecimals(6)
-                
-                peak_pos_form.addRow(f"Peak Position {peak_num}:", pp)
-                peak_pos_form.addRow(f"Weight Peak Position {peak_num}:", wpp)
-                
-                self.mass_target_spins[f"peak_position_{peak_num}_m{mass_num}"] = pp
-                self.mass_weight_spins[f"peak_position_{peak_num}_m{mass_num}"] = wpp
-            mg_layout.addWidget(peak_pos_group)
-
-            # Bandwidth group
-            bw_group = QGroupBox("Bandwidth Targets & Weights")
-            bw_form = QFormLayout(bw_group)
-            for i in range(1, 5):
-                for j in range(i+1, 5):
-                    bw = QDoubleSpinBox()
-                    bw.setRange(0, 1e6)
-                    bw.setDecimals(6)
-                    wbw = QDoubleSpinBox()
-                    wbw.setRange(0, 1e3)
-                    wbw.setDecimals(6)
-                    bw_form.addRow(f"Bandwidth {i}-{j}:", bw)
-                    bw_form.addRow(f"Weight Bandwidth {i}-{j}:", wbw)
-                    self.mass_target_spins[f"bandwidth_{i}_{j}_m{mass_num}"] = bw
-                    self.mass_weight_spins[f"bandwidth_{i}_{j}_m{mass_num}"] = wbw
-            mg_layout.addWidget(bw_group)
-
-            # Slope group
-            slope_group = QGroupBox("Slope Targets & Weights")
-            slope_form = QFormLayout(slope_group)
-            for i in range(1, 5):
-                for j in range(i+1, 5):
-                    s = QDoubleSpinBox()
-                    s.setRange(-1e6, 1e6)
-                    s.setDecimals(6)
-                    ws = QDoubleSpinBox()
-                    ws.setRange(0, 1e3)
-                    ws.setDecimals(6)
-                    slope_form.addRow(f"Slope {i}-{j}:", s)
-                    slope_form.addRow(f"Weight Slope {i}-{j}:", ws)
-                    self.mass_target_spins[f"slope_{i}_{j}_m{mass_num}"] = s
-                    self.mass_weight_spins[f"slope_{i}_{j}_m{mass_num}"] = ws
-            mg_layout.addWidget(slope_group)
-
-            # Area under curve group
-            auc_group = QGroupBox("Area Under Curve & Weight")
-            auc_form = QFormLayout(auc_group)
-            auc = QDoubleSpinBox()
-            auc.setRange(0, 1e6)
-            auc.setDecimals(6)
-            wauc = QDoubleSpinBox()
-            wauc.setRange(0, 1e3)
-            wauc.setDecimals(6)
-            auc_form.addRow("Area Under Curve:", auc)
-            auc_form.addRow("Weight Area Under Curve:", wauc)
-            self.mass_target_spins[f"area_under_curve_m{mass_num}"] = auc
-            self.mass_weight_spins[f"area_under_curve_m{mass_num}"] = wauc
-            mg_layout.addWidget(auc_group)
-
-            mg_layout.addStretch()
-            main_layout.addWidget(mass_group)
-        
-        main_layout.addStretch()
-        
-        # Set the container as the scroll area's widget
-        scroll_area.setWidget(main_container)
-        
-        # Add scroll area to the tab's layout
-        layout.addWidget(scroll_area)
-    
-    def create_omega_sensitivity_tab(self):
-        """Create the Omega points sensitivity analysis tab"""
-        self.omega_sensitivity_tab = QWidget()
-        layout = QVBoxLayout(self.omega_sensitivity_tab)
-
-        # Create tabs for parameters and visualization
-        self.sensitivity_tabs = ModernQTabWidget()
-        layout.addWidget(self.sensitivity_tabs)
-        
-        # --------- PARAMETERS TAB ---------
-        params_tab = QWidget()
-        params_layout = QVBoxLayout(params_tab)
-        
-        # Create a scroll area for parameters
-        params_scroll_area = QScrollArea()
-        params_scroll_area.setWidgetResizable(True)
-        
-        # Create main container widget for parameters
-        params_container = QWidget()
-        params_main_layout = QVBoxLayout(params_container)
-        
-        # Introduction group
-        intro_group = QGroupBox("Omega Points Sensitivity Analysis")
-        intro_layout = QVBoxLayout(intro_group)
-        
-        info_text = QLabel(
-            "This tool analyzes how the number of frequency points affects slope calculations in "
-            "the Frequency Response Function (FRF). It helps identify the minimum number of points "
-            "needed for stable results by incrementally increasing the frequency resolution and "
-            "observing the convergence of slope values."
-        )
-        info_text.setWordWrap(True)
-        intro_layout.addWidget(info_text)
-        
-        # Parameters group
-        params_group = QGroupBox("Analysis Parameters")
-        params_group_layout = QFormLayout(params_group)
-        
-        # Initial points
-        self.sensitivity_initial_points = QSpinBox()
-        self.sensitivity_initial_points.setRange(50, 5000)
-        self.sensitivity_initial_points.setValue(100)
-        params_group_layout.addRow("Initial Ω Points:", self.sensitivity_initial_points)
-        
-        # Maximum points
-        self.sensitivity_max_points = QSpinBox()
-        self.sensitivity_max_points.setRange(500, 1000000000)  # Allow very large values (10^9)
-        self.sensitivity_max_points.setValue(2000)
-        params_group_layout.addRow("Maximum Ω Points:", self.sensitivity_max_points)
-        
-        # Step size
-        self.sensitivity_step_size = QSpinBox()
-        self.sensitivity_step_size.setRange(10, 100000)  # Allow larger step sizes
-        self.sensitivity_step_size.setValue(1000)  # Increased default for large ranges
-        params_group_layout.addRow("Step Size:", self.sensitivity_step_size)
-        
-        # Convergence threshold
-        self.sensitivity_threshold = QDoubleSpinBox()
-        self.sensitivity_threshold.setRange(1e-10, 0.1)
-        self.sensitivity_threshold.setDecimals(10)
-        self.sensitivity_threshold.setSingleStep(1e-10)
-        self.sensitivity_threshold.setValue(0.01)
-        params_group_layout.addRow("Convergence Threshold:", self.sensitivity_threshold)
-        
-        # Max iterations
-        self.sensitivity_max_iterations = QSpinBox()
-        self.sensitivity_max_iterations.setRange(5, 1000000)  # Allow extremely high iteration counts
-        self.sensitivity_max_iterations.setValue(200)  # Set default to 200 to support larger ranges
-        params_group_layout.addRow("Maximum Iterations:", self.sensitivity_max_iterations)
-        
-        # Mass of interest
-        self.sensitivity_mass = QComboBox()
-        for i in range(1, 6):
-            self.sensitivity_mass.addItem(f"mass_{i}")
-        params_group_layout.addRow("Mass of Interest:", self.sensitivity_mass)
-        
-        # Plot results checkbox
-        self.sensitivity_plot_results = QCheckBox("Generate Convergence Plots")
-        self.sensitivity_plot_results.setChecked(True)
-        params_group_layout.addRow(self.sensitivity_plot_results)
-        
-        # Use optimal points checkbox
-        self.sensitivity_use_optimal = QCheckBox("Use Optimal Points in FRF Analysis")
-        self.sensitivity_use_optimal.setChecked(True)
-        params_group_layout.addRow(self.sensitivity_use_optimal)
-        
-        # Results group
-        self.sensitivity_results_group = QGroupBox("Analysis Results")
-        self.sensitivity_results_layout = QVBoxLayout(self.sensitivity_results_group)
-        
-        self.sensitivity_results_text = QTextEdit()
-        self.sensitivity_results_text.setReadOnly(True)
-        self.sensitivity_results_layout.addWidget(self.sensitivity_results_text)
-        
-        # Run button container
-        run_container = QWidget()
-        run_layout = QHBoxLayout(run_container)
-        run_layout.setContentsMargins(0, 20, 0, 0)  # Add some top margin
-        
-        # Add stretch to push button to center
-        run_layout.addStretch()
-        
-        # Create and style the Run Analysis button
-        self.run_sensitivity_btn = QPushButton("Run Sensitivity Analysis")
-        self.run_sensitivity_btn.setObjectName("primary-button")
-        self.run_sensitivity_btn.setMinimumWidth(200)
-        self.run_sensitivity_btn.setMinimumHeight(40)
-        self.run_sensitivity_btn.clicked.connect(self.run_omega_sensitivity)
-        self.run_sensitivity_btn.setStyleSheet("""
-            QPushButton#primary-button {
-                background-color: #4B67F0;
-                color: white;
-                border: none;
-                border-radius: 4px;
-                padding: 8px 20px;
-                font-weight: bold;
-                font-size: 14px;
-            }
-            QPushButton#primary-button:hover {
-                background-color: #3B57E0;
-            }
-            QPushButton#primary-button:pressed {
-                background-color: #2B47D0;
-            }
-        """)
-        run_layout.addWidget(self.run_sensitivity_btn)
-        
-        # Add stretch to push button to center
-        run_layout.addStretch()
-        
-        # Add all groups to main layout
-        params_main_layout.addWidget(intro_group)
-        params_main_layout.addWidget(params_group)
-        params_main_layout.addWidget(self.sensitivity_results_group)
-        params_main_layout.addWidget(run_container)
-        params_main_layout.addStretch()
-        
-        # Set the container as the scroll area's widget
-        params_scroll_area.setWidget(params_container)
-        params_layout.addWidget(params_scroll_area)
-        
-        # --------- VISUALIZATION TABS ---------
-        # Create visualization tabs widget
-        self.vis_tabs = ModernQTabWidget()
-        
-        # Common control panel for both visualization tabs
-        vis_control_panel = QWidget()
-        vis_control_layout = QHBoxLayout(vis_control_panel)
-        
-        # Save plot button
-        self.sensitivity_save_plot_btn = QPushButton("Save Current Plot")
-        self.sensitivity_save_plot_btn.setEnabled(False)
-        self.sensitivity_save_plot_btn.clicked.connect(self.save_sensitivity_plot)
-        vis_control_layout.addWidget(self.sensitivity_save_plot_btn)
-        
-        # Refresh plot button
-        self.sensitivity_refresh_plot_btn = QPushButton("Refresh Plots")
-        self.sensitivity_refresh_plot_btn.setEnabled(False)
-        self.sensitivity_refresh_plot_btn.clicked.connect(self.refresh_sensitivity_plot)
-        vis_control_layout.addWidget(self.sensitivity_refresh_plot_btn)
-        
-        # --------- CONVERGENCE PLOT TAB ---------
-        convergence_tab = QWidget()
-        convergence_layout = QVBoxLayout(convergence_tab)
-        
-        # Add control panel to layout
-        convergence_layout.addWidget(vis_control_panel)
-        
-        # Create figure canvas for convergence plot
-        self.convergence_fig = Figure(figsize=(10, 6))
-        self.convergence_canvas = FigureCanvas(self.convergence_fig)
-        self.convergence_canvas.setMinimumHeight(450)
-        self.convergence_toolbar = NavigationToolbar(self.convergence_canvas, convergence_tab)
-        
-        # Add canvas and toolbar to layout
-        convergence_layout.addWidget(self.convergence_canvas)
-        convergence_layout.addWidget(self.convergence_toolbar)
-        
-        # --------- RELATIVE CHANGE PLOT TAB ---------
-        rel_change_tab = QWidget()
-        rel_change_layout = QVBoxLayout(rel_change_tab)
-        
-        # Create figure canvas for relative change plot
-        self.rel_change_fig = Figure(figsize=(10, 6))
-        self.rel_change_canvas = FigureCanvas(self.rel_change_fig)
-        self.rel_change_canvas.setMinimumHeight(450)
-        self.rel_change_toolbar = NavigationToolbar(self.rel_change_canvas, rel_change_tab)
-        
-        # Add canvas and toolbar to layout
-        rel_change_layout.addWidget(self.rel_change_canvas)
-        rel_change_layout.addWidget(self.rel_change_toolbar)
-        
-        # No data message (added to both tabs)
-        self.convergence_no_data_label = QLabel("Run the sensitivity analysis to generate visualization")
-        self.convergence_no_data_label.setAlignment(Qt.AlignCenter)
-        self.convergence_no_data_label.setStyleSheet("color: #888; font-style: italic; font-size: 14px;")
-        convergence_layout.addWidget(self.convergence_no_data_label)
-        
-        self.rel_change_no_data_label = QLabel("Run the sensitivity analysis to generate visualization")
-        self.rel_change_no_data_label.setAlignment(Qt.AlignCenter)
-        self.rel_change_no_data_label.setStyleSheet("color: #888; font-style: italic; font-size: 14px;")
-        rel_change_layout.addWidget(self.rel_change_no_data_label)
-        
-        # Add tabs to the visualization tabs widget
-        self.vis_tabs.addTab(convergence_tab, "Slope Convergence")
-        self.vis_tabs.addTab(rel_change_tab, "Relative Change")
-        
-        # Create main visualization container tab
-        vis_tab = QWidget()
-        vis_layout = QVBoxLayout(vis_tab)
-        vis_layout.addWidget(self.vis_tabs)
-        
-        # Add tabs to the main tab widget
-        self.sensitivity_tabs.addTab(params_tab, "Parameters & Results")
-        self.sensitivity_tabs.addTab(vis_tab, "Visualization")
-        
-    def create_comparative_visualization_options(self, parent_layout):
-        """Create options for comparative visualization of multiple FRF inputs"""
-        comp_group = QGroupBox("Comparative Visualization")
-        comp_layout = QVBoxLayout(comp_group)
-        
-        # Introduction text
-        intro_label = QLabel("This section allows you to create custom comparative plots by selecting multiple FRF results and customizing legends and title.")
-        intro_label.setWordWrap(True)
-        comp_layout.addWidget(intro_label)
-        
-        # Available plots section
-        available_plots_group = QGroupBox("Available Plots")
-        available_plots_layout = QVBoxLayout(available_plots_group)
-        
-        self.available_plots_list = QListWidget()
-        self.available_plots_list.setSelectionMode(QAbstractItemView.MultiSelection)
-        self.available_plots_list.itemSelectionChanged.connect(self._update_legend_table_from_selection)
-        available_plots_layout.addWidget(self.available_plots_list)
-        
-        # Create button layout for management
-        plots_btn_layout = QHBoxLayout()
-        
-        # Add clear button for plot history
-        clear_plots_button = QPushButton("Clear All FRF Plots")
-        clear_plots_button.clicked.connect(self.clear_all_frf_plots)
-        plots_btn_layout.addWidget(clear_plots_button)
-        
-        # Add export/import buttons
-        export_button = QPushButton("Export FRF Data")
-        export_button.clicked.connect(self.export_frf_data)
-        plots_btn_layout.addWidget(export_button)
-        
-        import_button = QPushButton("Import FRF Data")
-        import_button.clicked.connect(self.import_frf_data)
-        plots_btn_layout.addWidget(import_button)
-        
-        available_plots_layout.addLayout(plots_btn_layout)
-        comp_layout.addWidget(available_plots_group)
-        
-        # Legend customization
-        legend_group = QGroupBox("Legend & Style Customization")
-        legend_layout = QVBoxLayout(legend_group)
-        
-        # Map of original plot names to custom properties
-        self.legend_map = {}
-        self.legend_table = QTableWidget()
-        self.legend_table.setColumnCount(5)
-        self.legend_table.setHorizontalHeaderLabels([
-            "Original Name", 
-            "Custom Legend", 
-            "Line Style", 
-            "Marker", 
-            "Color"
-        ])
-        self.legend_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        legend_layout.addWidget(self.legend_table)
-        
-        comp_layout.addWidget(legend_group)
-        
-        # Plot title customization
-        title_group = QGroupBox("Plot Title")
-        title_layout = QFormLayout(title_group)
-        
-        self.plot_title_edit = QLineEdit()
-        self.plot_title_edit.setPlaceholderText("Enter custom plot title here")
-        title_layout.addRow("Custom Title:", self.plot_title_edit)
-        
-        # Font size for title
-        self.title_font_size = QSpinBox()
-        self.title_font_size.setRange(8, 24)
-        self.title_font_size.setValue(14)
-        title_layout.addRow("Title Font Size:", self.title_font_size)
-        
-        comp_layout.addWidget(title_group)
-        
-        # Plot customization options
-        plot_options_group = QGroupBox("Plot Options")
-        plot_options_layout = QFormLayout(plot_options_group)
-        
-        # Figure size
-        fig_size_container = QWidget()
-        fig_size_layout = QHBoxLayout(fig_size_container)
-        fig_size_layout.setContentsMargins(0, 0, 0, 0)
-        
-        self.fig_width_spin = QSpinBox()
-        self.fig_width_spin.setRange(4, 20)
-        self.fig_width_spin.setValue(10)
-        fig_size_layout.addWidget(QLabel("Width:"))
-        fig_size_layout.addWidget(self.fig_width_spin)
-        
-        self.fig_height_spin = QSpinBox()
-        self.fig_height_spin.setRange(3, 15)
-        self.fig_height_spin.setValue(6)
-        fig_size_layout.addWidget(QLabel("Height:"))
-        fig_size_layout.addWidget(self.fig_height_spin)
-        
-        plot_options_layout.addRow("Figure Size:", fig_size_container)
-        
-        # Add normalization options
-        norm_container = QWidget()
-        norm_layout = QHBoxLayout(norm_container)
-        norm_layout.setContentsMargins(0, 0, 0, 0)
-        
-        # X axis normalization
-        self.x_norm_check = QCheckBox("X-Axis")
-        norm_layout.addWidget(self.x_norm_check)
-        
-        self.x_norm_value = QDoubleSpinBox()
-        self.x_norm_value.setRange(0.001, 1000000)
-        self.x_norm_value.setValue(1.0)
-        self.x_norm_value.setDecimals(3)
-        self.x_norm_value.setSingleStep(0.1)
-        self.x_norm_value.setEnabled(False)
-        norm_layout.addWidget(self.x_norm_value)
-        
-        self.x_norm_check.toggled.connect(self.x_norm_value.setEnabled)
-        
-        norm_layout.addSpacing(20)
-        
-        # Y axis normalization
-        self.y_norm_check = QCheckBox("Y-Axis")
-        norm_layout.addWidget(self.y_norm_check)
-        
-        self.y_norm_value = QDoubleSpinBox()
-        self.y_norm_value.setRange(0.001, 1000000)
-        self.y_norm_value.setValue(1.0)
-        self.y_norm_value.setDecimals(3)
-        self.y_norm_value.setSingleStep(0.1)
-        self.y_norm_value.setEnabled(False)
-        norm_layout.addWidget(self.y_norm_value)
-        
-        self.y_norm_check.toggled.connect(self.y_norm_value.setEnabled)
-        
-        plot_options_layout.addRow("Normalize by:", norm_container)
-        
-        # Grid options
-        self.show_grid_check = QCheckBox()
-        self.show_grid_check.setChecked(True)
-        plot_options_layout.addRow("Show Grid:", self.show_grid_check)
-        
-        # Legend position
-        self.legend_position_combo = QComboBox()
-        for pos in ["best", "upper right", "upper left", "lower left", "lower right", 
-                   "right", "center left", "center right", "lower center", "upper center", "center"]:
-            self.legend_position_combo.addItem(pos)
-        plot_options_layout.addRow("Legend Position:", self.legend_position_combo)
-        
-        comp_layout.addWidget(plot_options_group)
-        
-        # Visualization actions
-        actions_container = QWidget()
-        actions_layout = QHBoxLayout(actions_container)
-        
-        self.create_comp_plot_btn = QPushButton("Create Comparative Plot")
-        self.create_comp_plot_btn.setObjectName("primary-button")
-        self.create_comp_plot_btn.clicked.connect(self.create_comparative_plot)
-        actions_layout.addWidget(self.create_comp_plot_btn)
-        
-        self.save_comp_plot_btn = QPushButton("Save Plot")
-        self.save_comp_plot_btn.setObjectName("secondary-button")
-        self.save_comp_plot_btn.clicked.connect(lambda: self.save_plot(self.comp_fig, "Comparative FRF"))
-        actions_layout.addWidget(self.save_comp_plot_btn)
-        
-        comp_layout.addWidget(actions_container)
-        
-        parent_layout.addWidget(comp_group)
-        
-    def _update_legend_table_from_selection(self):
-        """Update the legend table based on the selected plots in the list widget"""
-        # Clear current table contents
-        self.legend_table.setRowCount(0)
-        
-        # Get selected items
-        selected_items = self.available_plots_list.selectedItems()
-        
-        if not selected_items:
-            return
-            
-        # Create a row for each selected plot
-        self.legend_table.setRowCount(len(selected_items))
-        
-        for row, item in enumerate(selected_items):
-            plot_name = item.text()
-            
-            # Original name column (non-editable)
-            name_item = QTableWidgetItem(plot_name)
-            name_item.setFlags(Qt.ItemIsEnabled)  # Make it non-editable
-            self.legend_table.setItem(row, 0, name_item)
-            
-            # Custom legend name column
-            if plot_name in self.legend_map and 'custom_name' in self.legend_map[plot_name]:
-                legend_name = self.legend_map[plot_name]['custom_name']
-            else:
-                legend_name = plot_name
-                
-            legend_item = QTableWidgetItem(legend_name)
-            self.legend_table.setItem(row, 1, legend_item)
-            
-            # Line style combo box
-            line_style_combo = QComboBox()
-            line_styles = ['-', '--', '-.', ':']
-            for style in line_styles:
-                line_style_combo.addItem(style)
-                
-            # Set previously selected style if available
-            if plot_name in self.legend_map and 'line_style' in self.legend_map[plot_name]:
-                line_style_value = self.legend_map[plot_name]['line_style']
-                # Convert empty string to "None" for the combobox
-                if line_style_value == "":
-                    line_style_value = "None"
-                try:
-                    index = line_styles.index(line_style_value)
-                    if index >= 0:
-                        line_style_combo.setCurrentIndex(index)
-                except ValueError:
-                    # If line style value not in the list, use default
-                    line_style_combo.setCurrentIndex(0)  # First style
-                    
-            self.legend_table.setCellWidget(row, 2, line_style_combo)
-            
-            # Marker combo box
-            marker_combo = QComboBox()
-            markers = ['None', '.', 'o', 'v', '^', '<', '>', 's', 'p', '*', 'h', 'H', '+', 'x', 'D', 'd', '|', '_']
-            for marker in markers:
-                marker_combo.addItem(marker)
-                
-            # Set previously selected marker if available
-            if plot_name in self.legend_map and 'marker' in self.legend_map[plot_name]:
-                marker_value = self.legend_map[plot_name]['marker']
-                # Convert empty string to "None" for the combobox
-                if marker_value == "":
-                    marker_value = "None"
-                try:
-                    index = markers.index(marker_value)
-                    if index >= 0:
-                        marker_combo.setCurrentIndex(index)
-                except ValueError:
-                    # If marker value not in the list, use default
-                    marker_combo.setCurrentIndex(0)  # "None"
-                    
-            self.legend_table.setCellWidget(row, 3, marker_combo)
-            
-            # Color button
-            color_button = QPushButton()
-            color_button.setAutoFillBackground(True)
-            
-            # Set previously selected color if available
-            if plot_name in self.legend_map and 'color' in self.legend_map[plot_name]:
-                color = self.legend_map[plot_name]['color']
-                color_button.setStyleSheet(f"background-color: {color};")
-            else:
-                # Generate random color if no previous color exists
-                import random
-                r, g, b = [random.randint(0, 255) for _ in range(3)]
-                color = f"rgb({r},{g},{b})"
-                color_button.setStyleSheet(f"background-color: {color};")
-            
-            # Connect button to color picker
-            color_button.clicked.connect(lambda checked, row=row: self._choose_color(row))
-            
-            self.legend_table.setCellWidget(row, 4, color_button)
-            
-            # Store initial values in the legend map
-            if plot_name not in self.legend_map:
-                self.legend_map[plot_name] = {
-                    'custom_name': legend_name,
-                    'line_style': line_styles[line_style_combo.currentIndex()],
-                    'marker': markers[marker_combo.currentIndex()],
-                    'color': color
-                }
-    
-    def _choose_color(self, row):
-        """Open a color dialog when a color button is clicked"""
-        from PyQt5.QtWidgets import QColorDialog
-        
-        color_button = self.legend_table.cellWidget(row, 4)
-        color_dialog = QColorDialog(self)
-        
-        if color_dialog.exec_():
-            color = color_dialog.selectedColor()
-            if color.isValid():
-                color_name = color.name()
-                color_button.setStyleSheet(f"background-color: {color_name};")
-                
-                # Update the legend map with the new color
-                plot_name = self.legend_table.item(row, 0).text()
-                if plot_name in self.legend_map:
-                    self.legend_map[plot_name]['color'] = color_name
-                else:
-                    self.legend_map[plot_name] = {'color': color_name}
-    
-    def clear_all_frf_plots(self):
-        """Clear all FRF plots from the list and reset the legend map"""
-        # Clear the list widget
-        self.available_plots_list.clear()
-        
-        # Clear the legend table
-        self.legend_table.setRowCount(0)
-        
-        # Reset the legend map
-        self.legend_map = {}
-        
-        # Clear any existing plot
-        try:
-            if hasattr(self, 'comp_fig') and self.comp_fig:
-                import matplotlib.pyplot as plt
-                plt.close(self.comp_fig)
-                self.comp_fig = None
-        except Exception as e:
-            print(f"Error clearing figures: {str(e)}")
-    
-    def export_frf_data(self):
-        """Export FRF data to a file"""
-        from PyQt5.QtWidgets import QFileDialog
-        import json
-        import os
-        
-        # Check if there are any plots to export
-        if self.available_plots_list.count() == 0:
-            QMessageBox.warning(self, "Export Error", "No FRF data available to export.")
-            return
-        
-        # Get export filename
-        filename, _ = QFileDialog.getSaveFileName(
-            self, "Export FRF Data", os.path.expanduser("~"), "JSON Files (*.json)"
-        )
-        
-        if not filename:
-            return  # User cancelled
-            
-        # Gather data to export
-        export_data = {
-            'plots': {},
-            'legend_map': self.legend_map
-        }
-        
-        # Add individual plot data
-        for i in range(self.available_plots_list.count()):
-            plot_name = self.available_plots_list.item(i).text()
-            if hasattr(self, f"frf_data_{plot_name}"):
-                plot_data = getattr(self, f"frf_data_{plot_name}")
-                export_data['plots'][plot_name] = plot_data
-        
-        # Write to file
-        try:
-            with open(filename, 'w') as f:
-                json.dump(export_data, f)
-            QMessageBox.information(self, "Export Successful", f"FRF data has been exported to {filename}")
-        except Exception as e:
-            QMessageBox.critical(self, "Export Error", f"Error exporting data: {str(e)}")
-    
-    def import_frf_data(self):
-        """Import FRF data from a file"""
-        from PyQt5.QtWidgets import QFileDialog
-        import json
-        import os
-        
-        # Get import filename
-        filename, _ = QFileDialog.getOpenFileName(
-            self, "Import FRF Data", os.path.expanduser("~"), "JSON Files (*.json)"
-        )
-        
-        if not filename:
-            return  # User cancelled
-            
-        # Read the file
-        try:
-            with open(filename, 'r') as f:
-                import_data = json.load(f)
-                
-            # Validate imported data
-            if 'plots' not in import_data or 'legend_map' not in import_data:
-                raise ValueError("Invalid file format: missing required data")
-                
-            # Update legend map
-            self.legend_map.update(import_data['legend_map'])
-            
-            # Import plot data
-            for plot_name, plot_data in import_data['plots'].items():
-                # Store the data
-                setattr(self, f"frf_data_{plot_name}", plot_data)
-                
-                # Add to list if not already there
-                found = False
-                for i in range(self.available_plots_list.count()):
-                    if self.available_plots_list.item(i).text() == plot_name:
-                        found = True
-                        break
-                        
-                if not found:
-                    self.available_plots_list.addItem(plot_name)
-                    
-            QMessageBox.information(self, "Import Successful", 
-                                   f"Imported {len(import_data['plots'])} FRF datasets from {filename}")
-                                   
-        except Exception as e:
-            QMessageBox.critical(self, "Import Error", f"Error importing data: {str(e)}")
-    
-    def create_sobol_analysis_tab(self):
-        self.sobol_tab = QWidget()
-        layout = QVBoxLayout(self.sobol_tab)
-
-        # Create sub-tabs widget
-        self.sobol_sub_tabs = QTabWidget()
-
-        # -------------------- Sub-tab 1: Sobol Analysis Settings --------------------
-        sobol_settings_tab = QWidget()
-        sobol_settings_layout = QFormLayout(sobol_settings_tab)
-
-        self.num_samples_line = QLineEdit()
-        self.num_samples_line.setPlaceholderText("e.g. 32,64,128")
-        sobol_settings_layout.addRow("Num Samples List:", self.num_samples_line)
-
-        self.n_jobs_spin = QSpinBox()
-        self.n_jobs_spin.setRange(1, 64)
-        self.n_jobs_spin.setValue(5)
-        sobol_settings_layout.addRow("Number of Jobs (n_jobs):", self.n_jobs_spin)
-
-        # Add a small Run Sobol button in the settings sub-tab
-        self.hyper_run_sobol_button = QPushButton("Run Sobol")
-        self.hyper_run_sobol_button.setFixedWidth(100)
-        self.hyper_run_sobol_button.clicked.connect(self._run_sobol_implementation)
-        sobol_settings_layout.addRow("Run Sobol:", self.hyper_run_sobol_button)
-
-        # -------------------- Sub-tab 2: DVA Parameters --------------------
-        dva_param_tab = QWidget()
-        dva_param_layout = QVBoxLayout(dva_param_tab)
-
-        self.dva_param_table = QTableWidget()
-        dva_parameters = [
-            *[f"beta_{i}" for i in range(1,16)],
-            *[f"lambda_{i}" for i in range(1,16)],
-            *[f"mu_{i}" for i in range(1,4)],
-            *[f"nu_{i}" for i in range(1,16)]
-        ]
-        self.dva_param_table.setRowCount(len(dva_parameters))
-        self.dva_param_table.setColumnCount(5)
-        self.dva_param_table.setHorizontalHeaderLabels(
-            ["Parameter", "Fixed", "Fixed Value", "Lower Bound", "Upper Bound"]
-        )
-        self.dva_param_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.dva_param_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-
-        for row, param in enumerate(dva_parameters):
-            param_item = QTableWidgetItem(param)
-            param_item.setFlags(Qt.ItemIsEnabled)
-            self.dva_param_table.setItem(row, 0, param_item)
-
-            fixed_checkbox = QCheckBox()
-            fixed_checkbox.stateChanged.connect(lambda state, r=row: self.toggle_fixed(state, r))
-            self.dva_param_table.setCellWidget(row, 1, fixed_checkbox)
-
-            fixed_value_spin = QDoubleSpinBox()
-            fixed_value_spin.setRange(-1e6, 1e6)
-            fixed_value_spin.setDecimals(6)
-            fixed_value_spin.setEnabled(False)
-            self.dva_param_table.setCellWidget(row, 2, fixed_value_spin)
-
-            lower_bound_spin = QDoubleSpinBox()
-            lower_bound_spin.setRange(-1e6, 1e6)
-            lower_bound_spin.setDecimals(6)
-            lower_bound_spin.setEnabled(True)
-            self.dva_param_table.setCellWidget(row, 3, lower_bound_spin)
-
-            upper_bound_spin = QDoubleSpinBox()
-            upper_bound_spin.setRange(-1e6, 1e6)
-            upper_bound_spin.setDecimals(6)
-            upper_bound_spin.setEnabled(True)
-            self.dva_param_table.setCellWidget(row, 4, upper_bound_spin)
-
-            # Default ranges
-            if param.startswith("beta_") or param.startswith("lambda_") or param.startswith("nu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(2.5)
-            elif param.startswith("mu_"):
-                lower_bound_spin.setValue(0.0001)
-                upper_bound_spin.setValue(0.75)
-            else:
-                lower_bound_spin.setValue(0.0)
-                upper_bound_spin.setValue(1.0)
-
-        dva_param_layout.addWidget(self.dva_param_table)
-
-        # -------------------- Sub-tab 3: Sobol Results --------------------
-        sobol_results_tab = QWidget()
-        sobol_results_layout = QVBoxLayout(sobol_results_tab)
-        
-        # Results text area
-        self.sobol_results_text = QTextEdit()
-        self.sobol_results_text.setReadOnly(True)
-        self.sobol_results_text.setStyleSheet("font-family: monospace;")
-        sobol_results_layout.addWidget(self.sobol_results_text)
-        
-        # Placeholder for Sobol plots
-        plot_container = QWidget()
-        plot_layout = QVBoxLayout(plot_container)
-        
-        # Plotting controls section
-        plot_controls = QWidget()
-        plot_controls_layout = QHBoxLayout(plot_controls)
-        
-        self.sobol_combo = QComboBox()
-        self.sobol_combo.currentIndexChanged.connect(self.update_sobol_plot)
-        plot_controls_layout.addWidget(QLabel("Plot Type:"))
-        plot_controls_layout.addWidget(self.sobol_combo)
-        
-        save_button = QPushButton("Save Results")
-        save_button.clicked.connect(self.save_sobol_results)
-        plot_controls_layout.addWidget(save_button)
-        
-        plot_layout.addWidget(plot_controls)
-        
-        # Figure canvas
-        self.sobol_figure = Figure(figsize=(4, 4))
-        self.sobol_canvas = FigureCanvas(self.sobol_figure)
-        self.sobol_canvas.setMinimumHeight(300)
-        toolbar = NavigationToolbar(self.sobol_canvas, self)
-        
-        plot_layout.addWidget(toolbar)
-        plot_layout.addWidget(self.sobol_canvas)
-        
-        sobol_results_layout.addWidget(plot_container)
-        
-        # Initialize plots dictionary
-        self.sobol_plots = {}
-
-        # Add sub-tabs to the Sobol tab widget
-        self.sobol_sub_tabs.addTab(sobol_settings_tab, "Analysis Settings")
-        self.sobol_sub_tabs.addTab(dva_param_tab, "DVA Parameters")
-        self.sobol_sub_tabs.addTab(sobol_results_tab, "Results")
-
-        # Add the Sobol sub-tabs widget to the main Sobol tab layout
-        layout.addWidget(self.sobol_sub_tabs)
-        self.sobol_tab.setLayout(layout)
-
-    def create_ga_tab(self):
-        """Create the genetic algorithm optimization tab"""
-        self.ga_tab = QWidget()
-        layout = QVBoxLayout(self.ga_tab)
-
-        # Create sub-tabs widget
-        self.ga_sub_tabs = QTabWidget()
-
-        # -------------------- Sub-tab 1: GA Hyperparameters --------------------
-        ga_hyper_tab = QWidget()
-        ga_hyper_layout = QFormLayout(ga_hyper_tab)
-
-        self.ga_pop_size_box = QSpinBox()
-        self.ga_pop_size_box.setRange(1, 10000)
-        self.ga_pop_size_box.setValue(800)
-
-        self.ga_num_generations_box = QSpinBox()
-        self.ga_num_generations_box.setRange(1, 10000)
-        self.ga_num_generations_box.setValue(100)
-
-        self.ga_cxpb_box = QDoubleSpinBox()
-        self.ga_cxpb_box.setRange(0, 1)
-        self.ga_cxpb_box.setValue(0.7)
-        self.ga_cxpb_box.setDecimals(3)
-
-        self.ga_mutpb_box = QDoubleSpinBox()
-        self.ga_mutpb_box.setRange(0, 1)
-        self.ga_mutpb_box.setValue(0.2)
-        self.ga_mutpb_box.setDecimals(3)
-
-        self.ga_tol_box = QDoubleSpinBox()
-        self.ga_tol_box.setRange(0, 1e6)
-        self.ga_tol_box.setValue(1e-3)
-        self.ga_tol_box.setDecimals(6)
-
-        self.ga_alpha_box = QDoubleSpinBox()
-        self.ga_alpha_box.setRange(0.0, 10.0)
-        self.ga_alpha_box.setDecimals(4)
-        self.ga_alpha_box.setSingleStep(0.01)
-        self.ga_alpha_box.setValue(0.01)
-        
-        # Add benchmarking runs box
-        self.ga_benchmark_runs_box = QSpinBox()
-        self.ga_benchmark_runs_box.setRange(1, 1000)
-        self.ga_benchmark_runs_box.setValue(1)
-        self.ga_benchmark_runs_box.setToolTip("Number of times to run the GA for benchmarking (1 = single run)")
-        
-        # Add adaptive rates checkbox
-        self.adaptive_rates_checkbox = QCheckBox("Use Adaptive Rates")
-        self.adaptive_rates_checkbox.setChecked(False)
-        self.adaptive_rates_checkbox.setToolTip("Automatically adjust crossover and mutation rates during optimization")
-        self.adaptive_rates_checkbox.stateChanged.connect(self.toggle_adaptive_rates_options)
-        
-        # Create a widget to hold adaptive rate options
-        self.adaptive_rates_options = QWidget()
-        adaptive_options_layout = QFormLayout(self.adaptive_rates_options)
-        adaptive_options_layout.setContentsMargins(20, 0, 0, 0)  # Add left margin for indentation
-        
-        # Stagnation limit spinner
-        self.stagnation_limit_box = QSpinBox()
-        self.stagnation_limit_box.setRange(1, 50)
-        self.stagnation_limit_box.setValue(5)
-        self.stagnation_limit_box.setToolTip("Number of generations without improvement before adapting rates")
-        adaptive_options_layout.addRow("Stagnation Limit:", self.stagnation_limit_box)
-        
-        # Create a widget for crossover bounds
-        crossover_bounds_widget = QWidget()
-        crossover_bounds_layout = QHBoxLayout(crossover_bounds_widget)
-        crossover_bounds_layout.setContentsMargins(0, 0, 0, 0)
-        
-        self.cxpb_min_box = QDoubleSpinBox()
-        self.cxpb_min_box.setRange(0.01, 0.5)
-        self.cxpb_min_box.setValue(0.1)
-        self.cxpb_min_box.setDecimals(2)
-        self.cxpb_min_box.setSingleStep(0.05)
-        self.cxpb_min_box.setToolTip("Minimum crossover probability")
-        
-        self.cxpb_max_box = QDoubleSpinBox()
-        self.cxpb_max_box.setRange(0.5, 1.0)
-        self.cxpb_max_box.setValue(0.9)
-        self.cxpb_max_box.setDecimals(2)
-        self.cxpb_max_box.setSingleStep(0.05)
-        self.cxpb_max_box.setToolTip("Maximum crossover probability")
-        
-        crossover_bounds_layout.addWidget(QLabel("Min:"))
-        crossover_bounds_layout.addWidget(self.cxpb_min_box)
-        crossover_bounds_layout.addWidget(QLabel("Max:"))
-        crossover_bounds_layout.addWidget(self.cxpb_max_box)
-        
-        adaptive_options_layout.addRow("Crossover Bounds:", crossover_bounds_widget)
-        
-        # Create a widget for mutation bounds
-        mutation_bounds_widget = QWidget()
-        mutation_bounds_layout = QHBoxLayout(mutation_bounds_widget)
-        mutation_bounds_layout.setContentsMargins(0, 0, 0, 0)
-        
-        self.mutpb_min_box = QDoubleSpinBox()
-        self.mutpb_min_box.setRange(0.01, 0.2)
-        self.mutpb_min_box.setValue(0.05)
-        self.mutpb_min_box.setDecimals(2)
-        self.mutpb_min_box.setSingleStep(0.01)
-        self.mutpb_min_box.setToolTip("Minimum mutation probability")
-        
-        self.mutpb_max_box = QDoubleSpinBox()
-        self.mutpb_max_box.setRange(0.2, 0.8)
-        self.mutpb_max_box.setValue(0.5)
-        self.mutpb_max_box.setDecimals(2)
-        self.mutpb_max_box.setSingleStep(0.05)
-        self.mutpb_max_box.setToolTip("Maximum mutation probability")
-        
-        mutation_bounds_layout.addWidget(QLabel("Min:"))
-        mutation_bounds_layout.addWidget(self.mutpb_min_box)
-        mutation_bounds_layout.addWidget(QLabel("Max:"))
-        mutation_bounds_layout.addWidget(self.mutpb_max_box)
-        
-        adaptive_options_layout.addRow("Mutation Bounds:", mutation_bounds_widget)
-        
-        # Initially hide adaptive options
-        self.adaptive_rates_options.setVisible(False)
-
-        ga_hyper_layout.addRow("Population Size:", self.ga_pop_size_box)
-        ga_hyper_layout.addRow("Number of Generations:", self.ga_num_generations_box)
-        ga_hyper_layout.addRow("Crossover Probability (cxpb):", self.ga_cxpb_box)
-        ga_hyper_layout.addRow("Mutation Probability (mutpb):", self.ga_mutpb_box)
-        ga_hyper_layout.addRow("Tolerance (tol):", self.ga_tol_box)
-        ga_hyper_layout.addRow("Sparsity Penalty (alpha):", self.ga_alpha_box)
-        ga_hyper_layout.addRow("Benchmark Runs:", self.ga_benchmark_runs_box)
-        ga_hyper_layout.addRow("", self.adaptive_rates_checkbox)
-        ga_hyper_layout.addRow("", self.adaptive_rates_options)
-
-        # Add a small Run GA button in the hyperparameters sub-tab
-        self.hyper_run_ga_button = QPushButton("Run GA")
-        self.hyper_run_ga_button.setFixedWidth(100)
-        self.hyper_run_ga_button.clicked.connect(self.run_ga)
-        ga_hyper_layout.addRow("Run GA:", self.hyper_run_ga_button)
-
-        # -------------------- Sub-tab 2: DVA Parameters --------------------
-        ga_param_tab = QWidget()
-        ga_param_layout = QVBoxLayout(ga_param_tab)
-
-        self.ga_param_table = QTableWidget()
-        dva_parameters = [
-            *[f"beta_{i}" for i in range(1,16)],
-            *[f"lambda_{i}" for i in range(1,16)],
-            *[f"mu_{i}" for i in range(1,4)],
-            *[f"nu_{i}" for i in range(1,16)]
-        ]
-        self.ga_param_table.setRowCount(len(dva_parameters))
-        self.ga_param_table.setColumnCount(5)
-        self.ga_param_table.setHorizontalHeaderLabels(
-            ["Parameter", "Fixed", "Fixed Value", "Lower Bound", "Upper Bound"]
-        )
-        self.ga_param_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.ga_param_table.setEditTriggers(QAbstractItemView.NoEditTriggers)
-
-        for row, param in enumerate(dva_parameters):
-            param_item = QTableWidgetItem(param)
-            param_item.setFlags(Qt.ItemIsEnabled)
-            self.ga_param_table.setItem(row, 0, param_item)
-
-            fixed_checkbox = QCheckBox()
-            fixed_checkbox.setChecked(True)  # Set fixed to true by default
-            fixed_checkbox.stateChanged.connect(lambda state, r=row: self.toggle_ga_fixed(state, r))
-            self.ga_param_table.setCellWidget(row, 1, fixed_checkbox)
-
-            fixed_value_spin = QDoubleSpinBox()
-            fixed_value_spin.setRange(0, 10e9)  # Changed to 0-10e9 range
-            fixed_value_spin.setDecimals(6)
-            fixed_value_spin.setValue(0.0)  # Set fixed value to 0
-            fixed_value_spin.setEnabled(True)  # Enable because fixed is checked
-            self.ga_param_table.setCellWidget(row, 2, fixed_value_spin)
-
-            lower_bound_spin = QDoubleSpinBox()
-            lower_bound_spin.setRange(0, 10e9)  # Changed to 0-10e9 range
-            lower_bound_spin.setDecimals(6)
-            lower_bound_spin.setValue(0.0)  # Set to 0
-            lower_bound_spin.setEnabled(False)  # Disable because fixed is checked
-            self.ga_param_table.setCellWidget(row, 3, lower_bound_spin)
-
-            upper_bound_spin = QDoubleSpinBox()
-            upper_bound_spin.setRange(0, 10e9)  # Changed to 0-10e9 range
-            upper_bound_spin.setDecimals(6)
-            upper_bound_spin.setValue(1.0)  # Set to 1
-            upper_bound_spin.setEnabled(False)  # Disable because fixed is checked
-            self.ga_param_table.setCellWidget(row, 4, upper_bound_spin)
-
-        ga_param_layout.addWidget(self.ga_param_table)
-
-        # -------------------- Sub-tab 3: Results --------------------
-        ga_results_tab = QWidget()
-        ga_results_layout = QVBoxLayout(ga_results_tab)
-
-        # Create a header area for label and export button
-        header_container = QWidget()
-        header_layout = QHBoxLayout(header_container)
-        header_layout.setContentsMargins(0, 0, 0, 0) # No margins for this internal layout
-
-        results_label = QLabel("GA Optimization Results:")
-        header_layout.addWidget(results_label)
-        header_layout.addStretch() # Add spacer to push the export button to the right
-
-        self.export_ga_results_button = QPushButton("Export GA Results")
-        self.export_ga_results_button.setObjectName("secondary-button") # Use existing styling if desired
-        self.export_ga_results_button.setToolTip("Export the GA optimization results to a JSON file")
-        self.export_ga_results_button.setEnabled(False)  # Initially disabled
-        # self.export_ga_results_button.clicked.connect(self.export_ga_results_to_file) # Will connect this later
-        header_layout.addWidget(self.export_ga_results_button)
-        
-        ga_results_layout.addWidget(header_container) # Add the header with label and button
-        
-        self.ga_results_text = QTextEdit()
-        self.ga_results_text.setReadOnly(True)
-        ga_results_layout.addWidget(self.ga_results_text)
-
-        # -------------------- Sub-tab 4: Benchmarking --------------------
-        ga_benchmark_tab = QWidget()
-        ga_benchmark_layout = QVBoxLayout(ga_benchmark_tab)
-
-        # Create buttons for import/export
-        button_container = QWidget()
-        button_layout = QHBoxLayout(button_container)
-        button_layout.setContentsMargins(0, 0, 0, 10)  # Add some bottom margin
-
-        self.import_benchmark_button = QPushButton("Import Benchmark Data")
-        self.import_benchmark_button.setToolTip("Import previously saved GA benchmark data")
-        self.import_benchmark_button.clicked.connect(self.import_ga_benchmark_data)
-        button_layout.addWidget(self.import_benchmark_button)
-
-        self.export_benchmark_button = QPushButton("Export Benchmark Data")
-        self.export_benchmark_button.setToolTip("Export current GA benchmark data to a file")
-        self.export_benchmark_button.setEnabled(False)  # Initially disabled until data is available
-        self.export_benchmark_button.clicked.connect(self.export_ga_benchmark_data)
-        button_layout.addWidget(self.export_benchmark_button)
-
-        button_layout.addStretch()  # Add stretch to push buttons to the left
-        ga_benchmark_layout.addWidget(button_container)
-
-        # Create tabs for different benchmark visualizations
-        self.benchmark_viz_tabs = QTabWidget()
-        
-        # Create tabs for different visualizations
-        violin_tab = QWidget()
-        violin_layout = QVBoxLayout(violin_tab)
-        self.violin_plot_widget = QWidget()
-        violin_layout.addWidget(self.violin_plot_widget)
-        
-        dist_tab = QWidget()
-        dist_layout = QVBoxLayout(dist_tab)
-        self.dist_plot_widget = QWidget()
-        dist_layout.addWidget(self.dist_plot_widget)
-        
-        scatter_tab = QWidget()
-        scatter_layout = QVBoxLayout(scatter_tab)
-        self.scatter_plot_widget = QWidget()
-        scatter_layout.addWidget(self.scatter_plot_widget)
-        
-        heatmap_tab = QWidget()
-        heatmap_layout = QVBoxLayout(heatmap_tab)
-        self.heatmap_plot_widget = QWidget()
-        heatmap_layout.addWidget(self.heatmap_plot_widget)
-        
-        # Add Q-Q plot tab
-        qq_tab = QWidget()
-        qq_layout = QVBoxLayout(qq_tab)
-        self.qq_plot_widget = QWidget()
-        qq_layout.addWidget(self.qq_plot_widget)
-        
-        # Summary statistics tabs (create subtabs for better organization)
-        stats_tab = QWidget()
-        stats_layout = QVBoxLayout(stats_tab)
-        
-        # Create a tabbed widget for the statistics section
-        stats_subtabs = QTabWidget()
-        
-        # ---- Subtab 1: Summary Statistics ----
-        summary_tab = QWidget()
-        summary_layout = QVBoxLayout(summary_tab)
-        
-        # Add summary statistics table
-        self.benchmark_stats_table = QTableWidget()
-        self.benchmark_stats_table.setColumnCount(5)
-        self.benchmark_stats_table.setHorizontalHeaderLabels(["Metric", "Min", "Max", "Mean", "Std"])
-        self.benchmark_stats_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        summary_layout.addWidget(QLabel("Statistical Summary of All Runs:"))
-        summary_layout.addWidget(self.benchmark_stats_table)
-        
-        # ---- Subtab 2: All Runs Table ----
-        runs_tab = QWidget()
-        runs_layout = QVBoxLayout(runs_tab)
-        
-        # Create a table for all runs
-        self.benchmark_runs_table = QTableWidget()
-        self.benchmark_runs_table.setColumnCount(4)
-        self.benchmark_runs_table.setHorizontalHeaderLabels(["Run #", "Fitness", "Rank", "Details"])
-        self.benchmark_runs_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.benchmark_runs_table.setSelectionBehavior(QAbstractItemView.SelectRows)
-        self.benchmark_runs_table.setSelectionMode(QAbstractItemView.SingleSelection)
-        self.benchmark_runs_table.itemClicked.connect(self.show_run_details)
-        
-        runs_layout.addWidget(QLabel("All Benchmark Runs:"))
-        runs_layout.addWidget(self.benchmark_runs_table)
-        
-        # Create run details text area
-        details_tab = QWidget()
-        details_layout = QVBoxLayout(details_tab)
-        details_group = QGroupBox("Run Details")
-        group_layout = QVBoxLayout(details_group)
-        self.run_details_text = QTextEdit()
-        self.run_details_text.setReadOnly(True)
-        group_layout.addWidget(self.run_details_text)
-        details_layout.addWidget(details_group)
-        
-        # Create GA Operations tab as a subtab
-        ga_ops_tab = QWidget()
-        ga_ops_layout = QVBoxLayout(ga_ops_tab)
-        self.ga_ops_plot_widget = QWidget()
-        ga_ops_layout.addWidget(self.ga_ops_plot_widget)
-        
-        # Add the subtabs to the stats tabbed widget
-        stats_subtabs.addTab(summary_tab, "Summary Statistics")
-        stats_subtabs.addTab(runs_tab, "All Runs")
-        stats_subtabs.addTab(details_tab, "Run Details")
-        stats_subtabs.addTab(ga_ops_tab, "GA Operations")
-        
-        # Add the stats tabbed widget to the stats tab
-        stats_layout.addWidget(stats_subtabs)
-        
-        # Add all visualization tabs to the benchmark visualization tabs
-        self.benchmark_viz_tabs.addTab(violin_tab, "Violin Plot")
-        self.benchmark_viz_tabs.addTab(dist_tab, "Distribution")
-        self.benchmark_viz_tabs.addTab(scatter_tab, "Scatter Plot")
-        self.benchmark_viz_tabs.addTab(heatmap_tab, "Parameter Correlations")
-        self.benchmark_viz_tabs.addTab(qq_tab, "Q-Q Plot")
-        self.benchmark_viz_tabs.addTab(stats_tab, "Statistics")
-        
-        # GA Operations Performance Tab - already added as a subtab of Statistics
-        
-        # Add the benchmark visualization tabs to the benchmark tab
-        ga_benchmark_layout.addWidget(self.benchmark_viz_tabs)
-        
-        # Add all sub-tabs to the GA tab widget
-        # Initialize empty benchmark data storage
-        self.ga_benchmark_data = []
-
-        # Add all sub-tabs to the GA tab widget
-        self.ga_sub_tabs.addTab(ga_hyper_tab, "GA Settings")
-        self.ga_sub_tabs.addTab(ga_param_tab, "DVA Parameters")
-        self.ga_sub_tabs.addTab(ga_results_tab, "Results")
-        self.ga_sub_tabs.addTab(ga_benchmark_tab, "GA Benchmarking")
-
-        # Add the GA sub-tabs widget to the main GA tab layout
-        layout.addWidget(self.ga_sub_tabs)
-        self.ga_tab.setLayout(layout)
-        
-    def toggle_fixed(self, state, row, table=None):
-        """Toggle the fixed state of a DVA parameter row"""
-        if table is None:
-            table = self.dva_param_table
-            
-        fixed = (state == Qt.Checked)
-        fixed_value_spin = table.cellWidget(row, 2)
-        lower_bound_spin = table.cellWidget(row, 3)
-        upper_bound_spin = table.cellWidget(row, 4)
-
-        fixed_value_spin.setEnabled(fixed)
-        lower_bound_spin.setEnabled(not fixed)
-        upper_bound_spin.setEnabled(not fixed)
-
-    def toggle_ga_fixed(self, state, row, table=None):
-        """Toggle the fixed state of a GA parameter row"""
-        if table is None:
-            table = self.ga_param_table
-            
-        fixed = (state == Qt.Checked)
-        fixed_value_spin = table.cellWidget(row, 2)
-        lower_bound_spin = table.cellWidget(row, 3)
-        upper_bound_spin = table.cellWidget(row, 4)
-        
-        # Enable/disable appropriate spinboxes
-        fixed_value_spin.setEnabled(fixed)
-        lower_bound_spin.setEnabled(not fixed)
-        upper_bound_spin.setEnabled(not fixed)
-        
-        # If switching to fixed mode, copy current lower bound value to fixed value
-        if fixed:
-            fixed_value_spin.setValue(lower_bound_spin.value())
-        # If switching to range mode, ensure lower bound is not greater than upper bound
-        else:
-            if lower_bound_spin.value() > upper_bound_spin.value():
-                upper_bound_spin.setValue(lower_bound_spin.value())
-
-    def toggle_adaptive_rates_options(self, state):
-        """Show or hide adaptive rates options based on checkbox state"""
-        self.adaptive_rates_options.setVisible(state == Qt.Checked)
-        
-        # Enable/disable the fixed rate inputs based on adaptive rates setting
-        self.ga_cxpb_box.setEnabled(state != Qt.Checked)
-        self.ga_mutpb_box.setEnabled(state != Qt.Checked)
-        
-        # Update tooltips to indicate that rates will be adaptive
-        if state == Qt.Checked:
-            self.ga_cxpb_box.setToolTip("Starting crossover probability (will adapt during optimization)")
-            self.ga_mutpb_box.setToolTip("Starting mutation probability (will adapt during optimization)")
-        else:
-            self.ga_cxpb_box.setToolTip("Crossover probability")
-            self.ga_mutpb_box.setToolTip("Mutation probability")
 
     def _run_sobol_implementation(self):
         """Run Sobol sensitivity analysis - main implementation"""
@@ -7397,318 +4767,6 @@ class MainWindow(QMainWindow):
         self.status_bar.showMessage("Reset to default values")
         # Reset logic for all parameters would be implemented here
 
-    def create_menubar(self):
-        """Create the application menubar with modern styling"""
-        menubar = self.menuBar()
-        
-        # File menu
-        file_menu = menubar.addMenu("&File")
-        
-        # New Project
-        new_action = QAction("&New Project", self)
-        new_action.setShortcut("Ctrl+N")
-        new_action.triggered.connect(lambda: self.status_bar.showMessage("New Project - Feature coming soon"))
-        file_menu.addAction(new_action)
-        
-        # Open Project
-        open_action = QAction("&Open Project", self)
-        open_action.setShortcut("Ctrl+O")
-        open_action.triggered.connect(lambda: self.status_bar.showMessage("Open Project - Feature coming soon"))
-        file_menu.addAction(open_action)
-        
-        file_menu.addSeparator()
-        
-        # Save Project
-        save_action = QAction("&Save Project", self)
-        save_action.setShortcut("Ctrl+S")
-        save_action.triggered.connect(lambda: self.status_bar.showMessage("Save Project - Feature coming soon"))
-        file_menu.addAction(save_action)
-        
-        # Save Project As
-        save_as_action = QAction("Save Project &As", self)
-        save_as_action.setShortcut("Ctrl+Shift+S")
-        save_as_action.triggered.connect(lambda: self.status_bar.showMessage("Save Project As - Feature coming soon"))
-        file_menu.addAction(save_as_action)
-        
-        file_menu.addSeparator()
-        
-        # Import
-        import_action = QAction("&Import Parameters", self)
-        import_action.triggered.connect(self.import_parameters)
-        file_menu.addAction(import_action)
-        
-        # Export
-        export_action = QAction("&Export Parameters", self)
-        export_action.triggered.connect(self.export_parameters)
-        file_menu.addAction(export_action)
-        
-        file_menu.addSeparator()
-        
-        # Exit
-        exit_action = QAction("E&xit", self)
-        exit_action.setShortcut("Alt+F4")
-        exit_action.triggered.connect(self.close)
-        file_menu.addAction(exit_action)
-        
-        # Edit menu
-        edit_menu = menubar.addMenu("&Edit")
-        
-        # Default values
-        default_action = QAction("Reset to &Default Values", self)
-        default_action.triggered.connect(self.set_default_values)
-        edit_menu.addAction(default_action)
-        
-        # View menu
-        view_menu = menubar.addMenu("&View")
-        
-        # Toggle Sidebar
-        sidebar_action = QAction("Toggle &Sidebar", self)
-        sidebar_action.setShortcut("Ctrl+B")
-        sidebar_action.triggered.connect(lambda: self.status_bar.showMessage("Toggle Sidebar - Feature coming soon"))
-        view_menu.addAction(sidebar_action)
-        
-        view_menu.addSeparator()
-        
-        # Theme submenu
-        theme_menu = view_menu.addMenu("&Theme")
-        
-        # Theme action group to make selections exclusive
-        theme_group = QActionGroup(self)
-        
-        # Dark theme action
-        dark_action = QAction("&Dark Theme", self)
-        dark_action.setCheckable(True)
-        if self.current_theme == 'Dark':
-            dark_action.setChecked(True)
-        dark_action.triggered.connect(lambda: self.switch_theme('Dark'))
-        theme_group.addAction(dark_action)
-        theme_menu.addAction(dark_action)
-        
-        # Light theme action
-        light_action = QAction("&Light Theme", self)
-        light_action.setCheckable(True)
-        if self.current_theme == 'Light':
-            light_action.setChecked(True)
-        light_action.triggered.connect(lambda: self.switch_theme('Light'))
-        theme_group.addAction(light_action)
-        theme_menu.addAction(light_action)
-        
-        # Tools menu
-        tools_menu = menubar.addMenu("&Tools")
-        
-        # Run FRF
-        run_frf_action = QAction("Run &FRF Analysis", self)
-        run_frf_action.triggered.connect(self.run_frf)
-        tools_menu.addAction(run_frf_action)
-        
-        # Run Sobol
-        run_sobol_action = QAction("Run &Sobol Analysis", self)
-        run_sobol_action.setToolTip("Run Sobol Analysis")
-        run_sobol_action.setVisible(False)  # Hide button
-        tools_menu.addAction(run_sobol_action)
-        
-        tools_menu.addSeparator()
-        
-        # Optimization submenu
-        optimization_menu = tools_menu.addMenu("&Optimization")
-        
-        # GA Optimization
-        ga_action = QAction("&Genetic Algorithm", self)
-        ga_action.triggered.connect(self.run_ga)
-        optimization_menu.addAction(ga_action)
-        
-        # PSO Optimization
-        pso_action = QAction("&Particle Swarm", self)
-        pso_action.triggered.connect(self.run_pso)
-        optimization_menu.addAction(pso_action)
-        
-        # DE Optimization
-        de_action = QAction("&Differential Evolution", self)
-        de_action.triggered.connect(self.run_de)
-        optimization_menu.addAction(de_action)
-        
-        # SA Optimization
-        sa_action = QAction("&Simulated Annealing", self)
-        sa_action.triggered.connect(self.run_sa)
-        optimization_menu.addAction(sa_action)
-        
-        # CMAES Optimization
-        cmaes_action = QAction("&CMA-ES", self)
-        cmaes_action.triggered.connect(self.run_cmaes)
-        optimization_menu.addAction(cmaes_action)
-        
-        # RL Optimization removed
-        
-        # Help menu
-        help_menu = menubar.addMenu("&Help")
-        
-        # Documentation
-        docs_action = QAction("&Documentation", self)
-        docs_action.triggered.connect(lambda: self.status_bar.showMessage("Documentation - Feature coming soon"))
-        help_menu.addAction(docs_action)
-        
-        # About
-        about_action = QAction("&About", self)
-        about_action.triggered.connect(self.show_about)
-        help_menu.addAction(about_action)
-    
-    def show_about(self):
-        """Show about dialog"""
-        QMessageBox.about(self, "About DeVana", 
-                          "DeVana v2.0\n\n"
-                          "A modern application for designing and optimizing vibration systems.\n\n"
-                          "© 2023 DeVana Team\n"
-                          "All rights reserved.")
-        
-    def create_toolbar(self):
-        """Create the application toolbar with modern styling"""
-        toolbar = QToolBar("Main Toolbar")
-        toolbar.setMovable(False)
-        toolbar.setIconSize(QSize(24, 24))
-        self.addToolBar(toolbar)
-        
-        # Add spacer at the beginning
-        spacer = QWidget()
-        spacer.setFixedWidth(10)
-        toolbar.addWidget(spacer)
-        
-        # New Project button
-        new_button = QPushButton("New Project")
-        new_button.setObjectName("toolbar-button")
-        new_button.setToolTip("Create a new project")
-        new_button.clicked.connect(lambda: self.status_bar.showMessage("New Project - Feature coming soon"))
-        toolbar.addWidget(new_button)
-        
-        # Open Project button
-        open_button = QPushButton("Open Project")
-        open_button.setObjectName("toolbar-button")
-        open_button.setToolTip("Open an existing project")
-        open_button.clicked.connect(lambda: self.status_bar.showMessage("Open Project - Feature coming soon"))
-        toolbar.addWidget(open_button)
-        
-        # Save Project button
-        save_button = QPushButton("Save Project")
-        save_button.setObjectName("toolbar-button")
-        save_button.setToolTip("Save the current project")
-        save_button.clicked.connect(lambda: self.status_bar.showMessage("Save Project - Feature coming soon"))
-        toolbar.addWidget(save_button)
-        
-        # Add separator
-        toolbar.addSeparator()
-        
-        # Run FRF button
-        run_frf_button = QPushButton("Run FRF")
-        run_frf_button.setObjectName("primary-button")
-        run_frf_button.setToolTip("Run FRF Analysis")
-        run_frf_button.clicked.connect(self.run_frf)
-        run_frf_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_frf_button)
-        
-        # Run Sobol button
-        run_sobol_button = QPushButton("Run Sobol")
-        run_sobol_button.setObjectName("primary-button")
-        run_sobol_button.setToolTip("Run Sobol Analysis")
-        run_sobol_button.clicked.connect(self._run_sobol_implementation)
-        run_sobol_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_sobol_button)
-        
-        # Run PSO button
-        run_pso_button = QPushButton("Run PSO")
-        run_pso_button.setObjectName("primary-button")
-        run_pso_button.setToolTip("Run Particle Swarm Optimization")
-        run_pso_button.clicked.connect(self.run_pso)
-        run_pso_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_pso_button)
-        
-        # Run DE button
-        run_de_button = QPushButton("Run DE")
-        run_de_button.setObjectName("primary-button")
-        run_de_button.setToolTip("Run Differential Evolution")
-        run_de_button.clicked.connect(self.run_de)
-        run_de_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_de_button)
-        
-        # Run SA button
-        run_sa_button = QPushButton("Run SA")
-        run_sa_button.setObjectName("primary-button")
-        run_sa_button.setToolTip("Run Simulated Annealing")
-        run_sa_button.clicked.connect(self.run_sa)
-        run_sa_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_sa_button)
-        
-        # Run CMA-ES button
-        run_cmaes_button = QPushButton("Run CMA-ES")
-        run_cmaes_button.setObjectName("primary-button")
-        run_cmaes_button.setToolTip("Run CMA-ES Optimization")
-        run_cmaes_button.clicked.connect(self.run_cmaes)
-        run_cmaes_button.setVisible(False)  # Hide button
-        toolbar.addWidget(run_cmaes_button)
-        
-        # Add separator
-        toolbar.addSeparator()
-        
-        # Theme toggle
-        theme_button = QPushButton("Toggle Theme")
-        theme_button.setObjectName("toolbar-button")
-        theme_button.setToolTip(f"Switch to {'Light' if self.current_theme == 'Dark' else 'Dark'} Theme")
-        theme_button.clicked.connect(self.toggle_theme)
-        toolbar.addWidget(theme_button)
-        
-        # Add expanding spacer
-        spacer = QWidget()
-        spacer.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Preferred)
-        toolbar.addWidget(spacer)
-        
-    def switch_theme(self, theme):
-        """Switch the application theme"""
-        self.current_theme = theme
-        if theme == 'Dark':
-            self.apply_dark_theme()
-        else:
-            self.apply_light_theme()
-        
-        # Update theme toggle button tooltip
-        for action in self.findChildren(QAction):
-            if action.text() == "Toggle &Theme":
-                action.setToolTip(f"Switch to {'Light' if theme == 'Dark' else 'Dark'} Theme")
-
-
-    def create_microchip_controller_page(self):
-        """Create the empty microchip controller page for future implementation"""
-        microchip_page = QWidget()
-        layout = QVBoxLayout(microchip_page)
-        
-        # Centered content
-        center_widget = QWidget()
-        center_layout = QVBoxLayout(center_widget)
-        center_layout.setAlignment(Qt.AlignCenter)
-        
-        # Icon or image placeholder
-        placeholder = QLabel()
-        placeholder.setPixmap(QPixmap("placeholder_image.png" if os.path.exists("placeholder_image.png") else ""))
-        placeholder.setAlignment(Qt.AlignCenter)
-        center_layout.addWidget(placeholder)
-        
-        # Title
-        title = QLabel("Microchip Controller")
-        title.setFont(QFont("Segoe UI", 24, QFont.Bold))
-        title.setAlignment(Qt.AlignCenter)
-        center_layout.addWidget(title)
-        
-        # Description
-        description = QLabel("This feature will provide interfaces for microchip-based vibration control systems.")
-        description.setFont(QFont("Segoe UI", 12))
-        description.setAlignment(Qt.AlignCenter)
-        center_layout.addWidget(description)
-        
-        # Coming soon label
-        coming_soon = QLabel("Coming Soon!")
-        coming_soon.setFont(QFont("Segoe UI", 14, QFont.Bold))
-        coming_soon.setAlignment(Qt.AlignCenter)
-        center_layout.addWidget(coming_soon)
-        
-        layout.addWidget(center_widget)
-        self.content_stack.addWidget(microchip_page)
 
     def run_ga(self):
         """Run genetic algorithm optimization"""
@@ -10338,1211 +7396,4 @@ class MainWindow(QMainWindow):
         
         # Add scroll area to the tab's layout
         layout.addWidget(scroll_area)
-
-    def create_continuous_beam_page(self):
-        """Create the continuous beam analysis page"""
-        if not BEAM_IMPORTS_SUCCESSFUL:
-            # Create placeholder page if imports failed
-            beam_page = QWidget()
-            layout = QVBoxLayout(beam_page)
-            
-            # Centered content
-            center_widget = QWidget()
-            center_layout = QVBoxLayout(center_widget)
-            center_layout.setAlignment(Qt.AlignCenter)
-            
-            # Error message
-            error_label = QLabel("Continuous Beam Module Not Available")
-            error_label.setFont(QFont("Segoe UI", 24, QFont.Bold))
-            error_label.setAlignment(Qt.AlignCenter)
-            center_layout.addWidget(error_label)
-            
-            description = QLabel("Please make sure the 'Continues beam' module is correctly installed.")
-            description.setFont(QFont("Segoe UI", 12))
-            description.setAlignment(Qt.AlignCenter)
-            center_layout.addWidget(description)
-            
-            layout.addWidget(center_widget)
-            self.content_stack.addWidget(beam_page)
-            return
-            
-        # Create tab container
-        beam_page = QWidget()
-        page_layout = QVBoxLayout(beam_page)
-        page_layout.setContentsMargins(10, 10, 10, 10)
-        page_layout.setSpacing(10)
-        
-        # Header - more compact
-        header = QWidget()
-        header_layout = QHBoxLayout(header)
-        header_layout.setContentsMargins(5, 5, 5, 5)
-        
-        title_container = QVBoxLayout()
-        title = QLabel("Continuous Beam Analysis")
-        title.setFont(QFont("Segoe UI", 16, QFont.Bold))
-        title_container.addWidget(title)
-        
-        description = QLabel("Analyze and optimize vibration in continuous beams")
-        description.setFont(QFont("Segoe UI", 10))
-        title_container.addWidget(description)
-        
-        header_layout.addLayout(title_container)
-        header_layout.addStretch()
-        
-        # Add header to page layout with minimal height
-        header.setMaximumHeight(70)
-        page_layout.addWidget(header)
-        
-        # Initialize the beam specific properties
-        self.layers = []
-        self.force_regions_manager = ForceRegionManager()
-        
-        # Add a default layer
-        default_layer = {
-            'height': 0.05,
-            'E': 210e9,  # Steel
-            'rho': 7800
-        }
-        self.layers.append(default_layer)
-        
-        # Pre-initialize canvases to avoid NoneType errors
-        self.beam_anim = None
-        self.beam_canvas = None
-        self.node_anim = None
-        self.node_canvas = None
-        
-        # Create tabs
-        self.beam_tabs = ModernQTabWidget()
-        
-        # Add tabs
-        self.init_beam_parameters_tab()
-        self.init_layers_tab()
-        self.init_loads_tab()
-        self.init_beam_results_tab()
-        
-        # Add tabs to the page
-        page_layout.addWidget(self.beam_tabs)
-        
-        # Create button area
-        button_container = QWidget()
-        button_layout = QHBoxLayout(button_container)
-        button_layout.setContentsMargins(0, 10, 0, 0)
-        
-        # Add run button
-        self.run_beam_button = QPushButton("▶ Run Simulation")
-        self.run_beam_button.clicked.connect(self.run_beam_simulation)
-        self.run_beam_button.setMinimumHeight(40)
-        self.run_beam_button.setStyleSheet("""
-            QPushButton {
-                background-color: #4CAF50;
-                color: white;
-                font-weight: bold;
-                border-radius: 4px;
-                padding: 8px 16px;
-            }
-            QPushButton:hover {
-                background-color: #45a049;
-            }
-            QPushButton:pressed {
-                background-color: #388E3C;
-            }
-        """)
-        
-        # Create spacer to push buttons to the right
-        button_layout.addStretch()
-        button_layout.addWidget(self.run_beam_button)
-        
-        # Add button container to main layout
-        page_layout.addWidget(button_container)
-        
-        # Add the page to the stack
-        self.content_stack.addWidget(beam_page)
-        
-    def init_beam_parameters_tab(self):
-        """Initialize the beam parameters tab"""
-        beam_tab = QWidget()
-        beam_layout = QVBoxLayout(beam_tab)
-        beam_layout.setContentsMargins(10, 10, 10, 10)
-        beam_layout.setSpacing(10)
-        
-        # Material properties group
-        material_group = QGroupBox("Material Properties")
-        material_layout = QFormLayout(material_group)
-        material_layout.setFieldGrowthPolicy(QFormLayout.AllNonFixedFieldsGrow)
-        material_layout.setContentsMargins(10, 20, 10, 10)
-        material_layout.setSpacing(10)
-        
-        # Young's modulus
-        self.young_modulus = QDoubleSpinBox()
-        self.young_modulus.setRange(1e9, 1000e9)
-        self.young_modulus.setValue(210e9)
-        self.young_modulus.setSuffix(" Pa")
-        self.young_modulus.setDecimals(2)
-        self.young_modulus.setStepType(QDoubleSpinBox.AdaptiveDecimalStepType)
-        material_layout.addRow("Young's Modulus:", self.young_modulus)
-        
-        # Density
-        self.density = QDoubleSpinBox()
-        self.density.setRange(100, 20000)
-        self.density.setValue(7800)
-        self.density.setSuffix(" kg/m³")
-        self.density.setDecimals(0)
-        material_layout.addRow("Density:", self.density)
-        
-        # Beam geometry group
-        geometry_group = QGroupBox("Beam Geometry")
-        geometry_layout = QFormLayout(geometry_group)
-        geometry_layout.setFieldGrowthPolicy(QFormLayout.AllNonFixedFieldsGrow)
-        geometry_layout.setContentsMargins(10, 20, 10, 10)
-        geometry_layout.setSpacing(10)
-        
-        # Length
-        self.beam_length = QDoubleSpinBox()
-        self.beam_length.setRange(0.1, 100)
-        self.beam_length.setValue(10.0)
-        self.beam_length.setSuffix(" m")
-        self.beam_length.setDecimals(2)
-        geometry_layout.addRow("Length:", self.beam_length)
-        
-        # Width
-        self.beam_width = QDoubleSpinBox()
-        self.beam_width.setRange(0.001, 10)
-        self.beam_width.setValue(0.2)
-        self.beam_width.setSuffix(" m")
-        self.beam_width.setDecimals(3)
-        geometry_layout.addRow("Width:", self.beam_width)
-        
-        # Height
-        self.beam_height = QDoubleSpinBox()
-        self.beam_height.setRange(0.001, 10)
-        self.beam_height.setValue(0.3)
-        self.beam_height.setSuffix(" m")
-        self.beam_height.setDecimals(3)
-        geometry_layout.addRow("Height:", self.beam_height)
-        
-        # Number of elements
-        self.num_elements = QSpinBox()
-        self.num_elements.setRange(2, 100)
-        self.num_elements.setValue(10)
-        geometry_layout.addRow("Number of Elements:", self.num_elements)
-        
-        # Spring stiffness
-        self.k_spring = QDoubleSpinBox()
-        self.k_spring.setRange(0, 1e9)
-        self.k_spring.setValue(1e5)
-        self.k_spring.setSuffix(" N/m")
-        self.k_spring.setDecimals(0)
-        self.k_spring.setStepType(QDoubleSpinBox.AdaptiveDecimalStepType)
-        geometry_layout.addRow("Tip Spring Stiffness:", self.k_spring)
-        
-        # Add groups to layout
-        beam_layout.addWidget(material_group)
-        beam_layout.addWidget(geometry_group)
-        beam_layout.addStretch()
-        
-        # Add to tabs
-        self.beam_tabs.addTab(beam_tab, "Beam Parameters")
-        
-    def init_layers_tab(self):
-        """Initialize the layers tab for composite beam"""
-        layers_tab = QWidget()
-        main_layout = QHBoxLayout(layers_tab)
-        main_layout.setContentsMargins(20, 20, 20, 20)
-        main_layout.setSpacing(15)
-        
-        # Left side - Cross-section visualization (takes 60% of width)
-        viz_container = QWidget()
-        viz_layout = QVBoxLayout(viz_container)
-        viz_layout.setContentsMargins(0, 0, 0, 0)
-        
-        # Add title and description
-        viz_title = QLabel("Cross-Section Visualization")
-        viz_title.setFont(QFont("Segoe UI", 12, QFont.Bold))
-        viz_layout.addWidget(viz_title)
-        
-        viz_desc = QLabel("Visual representation of the beam's cross-section with layers")
-        viz_desc.setStyleSheet("color: #666; font-style: italic;")
-        viz_layout.addWidget(viz_desc)
-        
-        # Cross-section visualization - make it larger
-        self.cross_section_visualizer = CrossSectionVisualizer()
-        self.cross_section_visualizer.setMinimumHeight(350)
-        viz_layout.addWidget(self.cross_section_visualizer, 1)  # stretch factor of 1
-        
-        # Add dimension information
-        dimension_label = QLabel("Total Height: 0.0 m")
-        dimension_label.setAlignment(Qt.AlignCenter)
-        self.dimension_label = dimension_label  # Store for later updates
-        viz_layout.addWidget(dimension_label)
-        
-        main_layout.addWidget(viz_container, 60)  # 60% of width
-        
-        # Right side - Layers table and controls
-        controls_container = QWidget()
-        controls_layout = QVBoxLayout(controls_container)
-        controls_layout.setContentsMargins(0, 0, 0, 0)
-        
-        # Add title
-        controls_title = QLabel("Layer Properties")
-        controls_title.setFont(QFont("Segoe UI", 12, QFont.Bold))
-        controls_layout.addWidget(controls_title)
-        
-        # Layers table
-        self.layers_table = QTableWidget()
-        self.layers_table.setColumnCount(3)
-        self.layers_table.setHorizontalHeaderLabels(["Height (m)", "Young's Modulus (Pa)", "Density (kg/m³)"])
-        self.layers_table.horizontalHeader().setSectionResizeMode(QHeaderView.Stretch)
-        self.layers_table.setSelectionBehavior(QAbstractItemView.SelectRows)
-        self.layers_table.setSelectionMode(QAbstractItemView.SingleSelection)
-        self.layers_table.setMinimumHeight(200)
-        controls_layout.addWidget(self.layers_table)
-        
-        # Layer buttons
-        btn_layout = QHBoxLayout()
-        
-        self.add_layer_btn = QPushButton("Add Layer")
-        self.add_layer_btn.clicked.connect(self.add_new_layer)
-        btn_layout.addWidget(self.add_layer_btn)
-        
-        self.edit_layer_btn = QPushButton("Edit Layer")
-        self.edit_layer_btn.clicked.connect(self.edit_layer)
-        self.edit_layer_btn.setEnabled(False)
-        btn_layout.addWidget(self.edit_layer_btn)
-        
-        self.remove_layer_btn = QPushButton("Remove Layer")
-        self.remove_layer_btn.clicked.connect(self.remove_layer)
-        self.remove_layer_btn.setEnabled(False)
-        btn_layout.addWidget(self.remove_layer_btn)
-        
-        # Connect selection change
-        self.layers_table.itemSelectionChanged.connect(self.update_layer_buttons)
-        
-        # Add button layout
-        controls_layout.addLayout(btn_layout)
-        
-        # Add help text
-        help_text = QLabel(
-            "Add layers to define a composite beam. The total height "
-            "will be the sum of all layer heights. Layers are stacked "
-            "from top to bottom as shown in the visualization."
-        )
-        help_text.setWordWrap(True)
-        help_text.setStyleSheet("color: #666; font-style: italic;")
-        controls_layout.addWidget(help_text)
-        
-        # Add a stretch to push everything up
-        controls_layout.addStretch()
-        
-        main_layout.addWidget(controls_container, 40)  # 40% of width
-        
-        # Add existing layers to the table
-        for layer in self.layers:
-            self.add_layer_to_table(layer)
-        
-        # Update the cross-section
-        self.update_cross_section()
-        
-        # Add to tabs
-        self.beam_tabs.addTab(layers_tab, "Layers")
-    
-    def update_layer_buttons(self):
-        """Enable or disable layer editing buttons based on selection"""
-        selected = len(self.layers_table.selectedItems()) > 0
-        self.edit_layer_btn.setEnabled(selected)
-        self.remove_layer_btn.setEnabled(selected)
-    
-    def add_new_layer(self):
-        """Open dialog to add a new layer"""
-        dialog = LayerDialog(self)
-        if dialog.exec_():
-            layer_data = dialog.get_layer_data()
-            
-            # Convert string expressions to callables if needed
-            for key in ['E', 'rho']:
-                if isinstance(layer_data.get(key), str):
-                    from Continues_beam.utils import parse_expression
-                    try:
-                        layer_data[key] = parse_expression(layer_data[key])
-                    except ValueError as e:
-                        QMessageBox.warning(self, "Expression Error", str(e))
-                        return
-            
-            # Add to layers list
-            self.layers.append(layer_data)
-            
-            # Add to table
-            self.add_layer_to_table(layer_data)
-            
-            # Update visualization
-            self.update_cross_section()
-    
-    def edit_layer(self):
-        """Edit the selected layer"""
-        selected_row = self.layers_table.currentRow()
-        if selected_row >= 0 and selected_row < len(self.layers):
-            # Get current layer data
-            layer_data = self.layers[selected_row]
-            
-            # Open dialog with current data
-            dialog = LayerDialog(self, layer_data)
-            if dialog.exec_():
-                new_layer_data = dialog.get_layer_data()
-                
-                # Convert string expressions to callables if needed
-                for key in ['E', 'rho']:
-                    if isinstance(new_layer_data.get(key), str):
-                        from Continues_beam.utils import parse_expression
-                        try:
-                            new_layer_data[key] = parse_expression(new_layer_data[key])
-                        except ValueError as e:
-                            QMessageBox.warning(self, "Expression Error", str(e))
-                            return
-                
-                # Update layers list
-                self.layers[selected_row] = new_layer_data
-                
-                # Update table display
-                self.layers_table.setRowCount(0)
-                for layer in self.layers:
-                    self.add_layer_to_table(layer)
-                
-                # Update visualization
-                self.update_cross_section()    
-    def remove_layer(self):
-        """Remove the selected layer"""
-        selected_row = self.layers_table.currentRow()
-        if selected_row >= 0 and selected_row < len(self.layers):
-            # Confirm deletion
-            reply = QMessageBox.question(
-                self, 
-                "Confirm Deletion",
-                f"Remove layer {selected_row + 1}?",
-                QMessageBox.Yes | QMessageBox.No,
-                QMessageBox.No
-            )
-            
-            if reply == QMessageBox.Yes:
-                # Remove from layers list
-                del self.layers[selected_row]
-                
-                # Update table display
-                self.layers_table.setRowCount(0)
-                for layer in self.layers:
-                    self.add_layer_to_table(layer)
-                
-                # Update visualization
-                self.update_cross_section()
-                
-    def add_layer_to_table(self, layer):
-        """Add a layer to the table display
-        
-        Args:
-            layer (dict): Layer data dictionary
-        """
-        row = self.layers_table.rowCount()
-        self.layers_table.insertRow(row)
-        
-        # Height column
-        height_item = QTableWidgetItem(f"{layer['height']:.4f}")
-        height_item.setTextAlignment(Qt.AlignCenter)
-        self.layers_table.setItem(row, 0, height_item)
-        
-        # Young's modulus column
-        if 'E_expr' in layer:
-            E_text = layer['E_expr']
-        else:
-            E_text = f"{layer['E']:.2e}"
-        E_item = QTableWidgetItem(E_text)
-        E_item.setTextAlignment(Qt.AlignCenter)
-        self.layers_table.setItem(row, 1, E_item)
-        
-        # Density column
-        if 'rho_expr' in layer:
-            rho_text = layer['rho_expr']
-        else:
-            rho_text = f"{layer['rho']:.2f}"
-        rho_item = QTableWidgetItem(rho_text)
-        rho_item.setTextAlignment(Qt.AlignCenter)
-        self.layers_table.setItem(row, 2, rho_item)
-        
-    def init_loads_tab(self):
-        """Initialize the loads & time tab for continuous beam analysis"""
-        loads_tab = QWidget()
-        layout = QVBoxLayout(loads_tab)
-        layout.setContentsMargins(20, 20, 20, 20)
-        layout.setSpacing(15)
-        
-        # Create panel for managing force regions
-        regions_group = QGroupBox("Force Regions")
-        regions_layout = QVBoxLayout(regions_group)
-        self.force_regions_panel = ForceRegionsPanel(self.force_regions_manager)
-        regions_layout.addWidget(self.force_regions_panel)
-        layout.addWidget(regions_group)
-        
-        # Time settings group
-        time_group = QGroupBox("Time Settings")
-        time_layout = QFormLayout(time_group)
-        time_group.setLayout(time_layout)
-        
-        # Start time
-        self.time_start_spin = QDoubleSpinBox()
-        self.time_start_spin.setRange(0, 1000)
-        self.time_start_spin.setValue(0.0)
-        self.time_start_spin.setSuffix(" s")
-        self.time_start_spin.setDecimals(2)
-        time_layout.addRow("Start Time:", self.time_start_spin)
-        
-        # End time
-        self.time_end_spin = QDoubleSpinBox()
-        self.time_end_spin.setRange(0.01, 1000)
-        self.time_end_spin.setValue(3.0)
-        self.time_end_spin.setSuffix(" s")
-        self.time_end_spin.setDecimals(2)
-        time_layout.addRow("End Time:", self.time_end_spin)
-        
-        # Number of time points
-        self.time_points_spin = QSpinBox()
-        self.time_points_spin.setRange(10, 10000)
-        self.time_points_spin.setValue(300)
-        time_layout.addRow("Number of Time Points:", self.time_points_spin)
-        
-        layout.addWidget(time_group)
-        
-        # Stretch at the end to push everything to the top
-        layout.addStretch()
-        
-        # Add to tabs
-        self.beam_tabs.addTab(loads_tab, "Loads & Time")
-        
-    def update_cross_section(self):
-        """Update the cross-section visualizer with current layers"""
-        beam_width = self.beam_width.value()
-        self.cross_section_visualizer.set_layers(self.layers, beam_width)
-        
-        # Calculate the total height of all layers
-        total_height = sum(layer.get('height', 0.0) for layer in self.layers) if self.layers else 0.0
-        
-        # Update the dimension label with the total height
-        self.dimension_label.setText(f"Total Height: {total_height:.4f} m")
-        
-        # Ensure the visualizer repaints itself
-        self.cross_section_visualizer.update()
-
-    def init_beam_results_tab(self):
-        """Initialize the results tab for continuous beam analysis"""
-        # Create the new comprehensive results dashboard
-        try:
-            # Try to import from the new location first
-            try:
-                from src.ui.components.results_dashboard import ResultsDashboard
-            except ImportError:
-                # Fall back to the old location
-                from Continues_beam.ui.results_dashboard import ResultsDashboard
-            
-            results_tab = QWidget()
-            layout = QVBoxLayout(results_tab)
-            layout.setContentsMargins(0, 0, 0, 0)
-            layout.setSpacing(0)
-            
-            # Add saving capability header
-            header_layout = QHBoxLayout()
-            header_layout.setContentsMargins(10, 5, 10, 5)
-            
-            # Add save results button in the header
-            self.save_results_btn = QPushButton("💾 Save Results")
-            self.save_results_btn.clicked.connect(self.save_beam_results)
-            self.save_results_btn.setEnabled(False)  # Disabled until results are available
-            header_layout.addStretch()
-            header_layout.addWidget(self.save_results_btn)
-            
-            layout.addLayout(header_layout)
-            
-            # Create the dashboard
-            self.results_dashboard = ResultsDashboard()
-            
-            # Set size policy to allow the dashboard to expand
-            self.results_dashboard.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
-            
-            # Add the dashboard to the layout with a stretch factor
-            layout.addWidget(self.results_dashboard, 1)  # Give it a stretch factor of 1
-            
-            # For backward compatibility, maintain the frequency text widget
-            self.freq_text = QTextEdit()
-            self.freq_text.setVisible(False)  # Hide it as it's not needed in the new UI
-            
-            # Add to tabs
-            self.beam_tabs.addTab(results_tab, "Results")
-            
-        except ImportError as e:
-            # Fallback to old implementation if ResultsDashboard is not available
-            print(f"Warning: Could not import ResultsDashboard: {e}")
-            print("Using legacy results display instead.")
-            
-            results_tab = QWidget()
-            layout = QVBoxLayout(results_tab)
-            layout.setContentsMargins(5, 5, 5, 5)
-            layout.setSpacing(5)
-            
-            # Results display area
-            results_group = QGroupBox("Analysis Results")
-            results_layout = QVBoxLayout(results_group)
-            results_layout.setContentsMargins(5, 10, 5, 5)
-            results_layout.setSpacing(5)
-            
-            # Add natural frequencies display in a compact form
-            freq_header_layout = QHBoxLayout()
-            freq_label = QLabel("Natural Frequencies:")
-            freq_label.setFont(QFont("Segoe UI", 10, QFont.Bold))
-            freq_header_layout.addWidget(freq_label)
-            freq_header_layout.addStretch()
-            
-            # Add save results button in the header
-            self.save_results_btn = QPushButton("💾 Save Results")
-            self.save_results_btn.clicked.connect(self.save_beam_results)
-            self.save_results_btn.setEnabled(False)  # Disabled until results are available
-            freq_header_layout.addWidget(self.save_results_btn)
-            
-            results_layout.addLayout(freq_header_layout)
-            
-            self.freq_text = QTextEdit()
-            self.freq_text.setReadOnly(True)
-            self.freq_text.setMaximumHeight(80)  # Reduce height to save space
-            results_layout.addWidget(self.freq_text)
-            
-            # Create a splitter for the visualization area
-            viz_splitter = QSplitter(Qt.Horizontal)
-            viz_splitter.setChildrenCollapsible(False)
-            
-            # Left side - Static visualizations
-            left_viz = QTabWidget()
-            left_viz.setDocumentMode(True)
-            
-            # Beam deflection tab
-            beam_tab = QWidget()
-            beam_layout = QVBoxLayout(beam_tab)
-            beam_layout.setContentsMargins(2, 2, 2, 2)
-            self.beam_canvas = PlotCanvas(beam_tab)
-            beam_layout.addWidget(self.beam_canvas)
-            
-            # Node displacement over time tab
-            node_tab = QWidget()
-            node_layout = QVBoxLayout(node_tab)
-            node_layout.setContentsMargins(2, 2, 2, 2)
-            
-            # Node selection
-            node_select_layout = QHBoxLayout()
-            node_select_layout.addWidget(QLabel("Select Node:"))
-            
-            self.node_combo = QComboBox()
-            node_select_layout.addWidget(self.node_combo)
-            node_select_layout.addStretch()
-            
-            node_layout.addLayout(node_select_layout)
-            
-            self.node_canvas = PlotCanvas(node_tab)
-            node_layout.addWidget(self.node_canvas)
-            
-            # Add tabs to left visualization
-            left_viz.addTab(beam_tab, "Beam Deflection")
-            left_viz.addTab(node_tab, "Node Displacement")
-            
-            # Right side - Animations
-            right_viz = QTabWidget()
-            right_viz.setDocumentMode(True)
-            
-            # Beam animation tab
-            beam_anim_tab = QWidget()
-            beam_anim_layout = QVBoxLayout(beam_anim_tab)
-            beam_anim_layout.setContentsMargins(2, 2, 2, 2)
-            self.beam_animation_adapter = BeamAnimationAdapter()
-            beam_anim_layout.addWidget(self.beam_animation_adapter)
-            
-            # Mode shape animation tab
-            mode_shape_tab = QWidget()
-            mode_shape_layout = QVBoxLayout(mode_shape_tab)
-            mode_shape_layout.setContentsMargins(2, 2, 2, 2)
-            self.mode_shape_adapter = ModeShapeAdapter()
-            mode_shape_layout.addWidget(self.mode_shape_adapter)
-            
-            # Add tabs to right visualization
-            right_viz.addTab(beam_anim_tab, "Beam Animation")
-            right_viz.addTab(mode_shape_tab, "Mode Shapes")
-            
-            # Add both sides to the splitter
-            viz_splitter.addWidget(left_viz)
-            viz_splitter.addWidget(right_viz)
-            
-            # Set initial sizes to make them equal
-            viz_splitter.setSizes([500, 500])
-            
-            # Make the visualization area taller
-            viz_splitter.setMinimumHeight(1000)
-            
-            # Add the splitter to the results layout with stretch factor
-            results_layout.addWidget(viz_splitter, 1)  # Give it a stretch factor of 1
-            
-            # Add the results group to the main layout with stretch factor
-            layout.addWidget(results_group, 1)  # Give it a stretch factor of 1
-            
-            # Add to tabs
-            self.beam_tabs.addTab(results_tab, "Results")
-        
-    def run_beam_simulation(self):
-        """Run the beam vibration simulation"""
-        try:
-            # Update status
-            self.status_bar.showMessage("Running simulation...")
-            
-            # Get beam parameters
-            beam_length = self.beam_length.value()
-            beam_width = self.beam_width.value()
-            spring_constant = self.k_spring.value()
-            num_elements = self.num_elements.value()
-            
-            # Validate inputs - ensure we have at least one layer
-            if not self.layers or all(layer.get('height', 0) <= 0 for layer in self.layers):
-                QMessageBox.warning(self, "Simulation Error", 
-                                   "You must define at least one layer with a positive height.")
-                self.statusBar().showMessage("Simulation failed: No valid layers defined")
-                return
-            
-            # Prepare layers for solver
-            layers = []
-            for layer in self.layers:
-                height = layer.get('height', 0)
-                E = layer.get('E', 0)
-                rho = layer.get('rho', 0)
-                
-                layers.append({'height': height, 'E': E, 'rho': rho})
-            
-            # Set up force function
-            force_gens = get_force_generators()
-            
-            # Check if we have any defined force regions
-            if not self.force_regions_manager.regions:
-                msg = QMessageBox()
-                msg.setIcon(QMessageBox.Information)
-                msg.setWindowTitle("No Force Regions")
-                msg.setText("No force regions are defined.")
-                msg.setInformativeText("The simulation will run with a zero force function.\n"
-                                       "Consider adding force regions in the Loads & Time tab for meaningful results.")
-                msg.setStandardButtons(QMessageBox.Ok | QMessageBox.Cancel)
-                response = msg.exec_()
-                
-                if response == QMessageBox.Cancel:
-                    self.statusBar().showMessage("Simulation cancelled")
-                    return
-                
-                # Create a lambda function that always returns 0.0 as the force profile
-                f_profile = lambda x, t: 0.0
-            else:
-                # Create force profile from defined regions
-                f_profile = self.force_regions_manager.create_force_function(force_gens)
-            
-            # Run simulation
-            results = solve_beam_vibration(
-                width=beam_width,
-                layers=layers,
-                L=beam_length,
-                k_spring=spring_constant,
-                num_elems=num_elements,
-                f_profile=f_profile
-            )
-            
-            # Store results
-            self.simulation_results = results
-            
-            # Update display
-            self.update_results_display()
-            
-            # Update status
-            self.status_bar.showMessage("Simulation completed successfully")
-            
-        except Exception as e:
-            error_message = str(e)
-            detailed_message = "An error occurred during the simulation."
-            
-            # Check for common errors
-            if "no object chosen" in error_message.lower():
-                detailed_message = "A reference to an undefined object was detected. "
-                detailed_message += "This may happen if no force regions are defined or if there's an issue with the layers configuration."
-            
-            # Show error message
-            msg_box = QMessageBox()
-            msg_box.setIcon(QMessageBox.Critical)
-            msg_box.setWindowTitle("Simulation Error")
-            msg_box.setText(detailed_message)
-            msg_box.setDetailedText(f"Error details: {error_message}")
-            msg_box.exec_()
-            
-            # Update status bar
-            self.status_bar.showMessage(f"Simulation failed: {error_message}")
-    
-    def update_results_display(self):
-        """Update the results display with simulation results"""
-        if not hasattr(self, 'simulation_results') or not self.simulation_results:
-            return
-        
-        # Check if we're using the new results dashboard
-        if hasattr(self, 'results_dashboard'):
-            # Update the comprehensive dashboard with all results
-            self.results_dashboard.update_results(self.simulation_results)
-            
-            # Ensure the freq_text is also populated for backward compatibility
-            self.freq_text.clear()
-            for i, freq in enumerate(self.simulation_results['natural_frequencies_hz']):
-                if i < 10:  # Show only first 10 modes
-                    self.freq_text.append(f"Mode {i+1}: {freq:.2f} Hz")
-        else:
-            # Use the legacy display method
-            # Display natural frequencies
-            self.freq_text.clear()
-            for i, freq in enumerate(self.simulation_results['natural_frequencies_hz']):
-                if i < 10:  # Show only first 10 modes
-                    self.freq_text.append(f"Mode {i+1}: {freq:.2f} Hz")
-            
-            # Update beam deflection plot
-            self.plot_beam_deflection()
-            
-            # Update node selection combo
-            self.node_combo.clear()
-            for i in range(len(self.simulation_results['coords'])):
-                x = self.simulation_results['coords'][i]
-                self.node_combo.addItem(f"Node {i+1} (x={x:.2f}m)")
-            
-            # Connect node selection to plot update
-            self.node_combo.currentIndexChanged.connect(self.plot_node_displacement)
-            
-            # Initial node plot
-            if self.node_combo.count() > 0:
-                self.node_combo.setCurrentIndex(0)
-                self.plot_node_displacement()
-                
-            # Update animations
-            if hasattr(self, 'beam_animation_adapter'):
-                self.beam_animation_adapter.update_animation(self.simulation_results)
-                
-            if hasattr(self, 'mode_shape_adapter'):
-                self.mode_shape_adapter.update_results(self.simulation_results)
-            
-        # Enable save results button
-        self.save_results_btn.setEnabled(True)
-        
-    def plot_beam_deflection(self):
-        """Plot the beam deflection"""
-        if not hasattr(self, 'simulation_results') or not self.simulation_results:
-            return
-        
-        # Check if canvas exists
-        if not hasattr(self, 'beam_canvas') or self.beam_canvas is None:
-            return
-        
-        # Clear the canvas
-        self.beam_canvas.clear()
-        ax = self.beam_canvas.figure.add_subplot(111)
-        
-        # Get data
-        x = self.simulation_results['coords']
-        t = self.simulation_results['time']
-        u = self.simulation_results['displacement']
-        
-        # Plot beam deflection at selected time points
-        num_frames = min(10, len(t))
-        step = len(t) // num_frames
-        
-        # Extract displacement at even nodes (translations)
-        u_disp = u[::2, :]
-        
-        for i in range(0, len(t), step):
-            if i >= len(t):
-                continue
-            
-            # Get displacement at this time
-            deflection = u_disp[:, i]
-            
-            # Scale factor for visualization
-            scale = 1.0
-            if np.max(np.abs(deflection)) > 0:
-                scale = 0.2 * np.max(x) / np.max(np.abs(deflection))
-            
-            # Plot the deflected beam
-            ax.plot(x, scale * deflection, label=f"t={t[i]:.2f}s")
-        
-        # Plot settings
-        ax.set_xlabel('Position (m)')
-        ax.set_ylabel('Deflection (scaled)')
-        ax.set_title('Beam Deflection')
-        ax.grid(True)
-        ax.legend()
-        
-        # Update canvas
-        self.beam_canvas.draw()
-        
-    def plot_node_displacement(self):
-        """Plot the displacement of the selected node over time"""
-        if not hasattr(self, 'simulation_results') or not self.simulation_results:
-            return
-        
-        # Check if canvas exists
-        if not hasattr(self, 'node_canvas') or self.node_canvas is None:
-            return
-        
-        # Get selected node
-        if self.node_combo.count() == 0:
-            return
-        
-        node_idx = self.node_combo.currentIndex()
-        
-        # Clear the canvas
-        self.node_canvas.clear()
-        ax = self.node_canvas.figure.add_subplot(111)
-        
-        # Get data
-        t = self.simulation_results['time']
-        u = self.simulation_results['displacement']
-        v = self.simulation_results['velocity']
-        a = self.simulation_results['acceleration']
-        
-        # Get the displacement DOF for this node
-        dof = node_idx * 2  # 2 DOFs per node (displacement and rotation)
-        
-        # Plot displacement, velocity, and acceleration
-        ax.plot(t, u[dof, :], label='Displacement')
-        ax.plot(t, v[dof, :], label='Velocity')
-        ax.plot(t, a[dof, :] / 1000, label='Acceleration (÷1000)')
-        
-        # Plot settings
-        ax.set_xlabel('Time (s)')
-        ax.set_ylabel('Amplitude')
-        ax.set_title(f'Node {node_idx+1} Response')
-        ax.grid(True)
-        ax.legend()
-        
-        # Update canvas
-        self.node_canvas.draw()
-        
-    def save_beam_results(self):
-        """Save beam simulation results to CSV files"""
-        if not hasattr(self, 'simulation_results') or not self.simulation_results:
-            QMessageBox.warning(self, "No Results", "Please run a simulation before saving results.")
-            return
-            
-        try:
-            # Ask for directory to save results
-            save_dir = QFileDialog.getExistingDirectory(self, "Select Directory to Save Results")
-            if not save_dir:
-                return
-                
-            # Save natural frequencies
-            freq_file = os.path.join(save_dir, "natural_frequencies.csv")
-            with open(freq_file, 'w') as f:
-                f.write("Mode,Frequency (Hz)\n")
-                for i, freq in enumerate(self.simulation_results['natural_frequencies_hz']):
-                    f.write(f"{i+1},{freq:.6f}\n")
-                    
-            # Save displacement results
-            time = self.simulation_results['time']
-            disp = self.simulation_results['displacement']
-            coords = self.simulation_results['coords']
-            
-            # Create a DataFrame for displacement over time
-            disp_data = pd.DataFrame()
-            disp_data['Time (s)'] = time
-            
-            # Add columns for each node's displacement
-            for i in range(len(coords)):
-                dof = i * 2  # 2 DOFs per node (displacement and rotation)
-                disp_data[f'Node {i+1} (x={coords[i]:.2f}m)'] = disp[dof, :]
-                
-            # Save to CSV
-            disp_file = os.path.join(save_dir, "displacement_results.csv")
-            disp_data.to_csv(disp_file, index=False)
-            
-            # Create a DataFrame for the currently selected node's detailed response
-            if self.node_combo.count() > 0:
-                node_idx = self.node_combo.currentIndex()
-                dof = node_idx * 2
-                
-                node_data = pd.DataFrame()
-                node_data['Time (s)'] = time
-                node_data['Displacement'] = self.simulation_results['displacement'][dof, :]
-                node_data['Velocity'] = self.simulation_results['velocity'][dof, :]
-                node_data['Acceleration'] = self.simulation_results['acceleration'][dof, :]
-                
-                # Save to CSV
-                node_file = os.path.join(save_dir, f"node_{node_idx+1}_response.csv")
-                node_data.to_csv(node_file, index=False)
-            
-            # Show success message
-            QMessageBox.information(self, "Results Saved", 
-                                   f"Results successfully saved to {save_dir}")
-            
-        except Exception as e:
-            QMessageBox.critical(self, "Error Saving Results", f"Error: {str(e)}")
-
-    def format_parameter_name(self, param):
-        """
-        Format parameter names using LaTeX symbols for plotting.
-
-        Parameters:
-            param (str): The parameter name.
-
-        Returns:
-            str: Formatted parameter name.
-        """
-        GREEK_LETTERS = {
-            'beta': r'\beta',
-            'lambda': r'\lambda',
-            'mu': r'\mu',
-            'nu': r'\nu'
-        }
-        for greek_letter, symbol in GREEK_LETTERS.items():
-            if param.startswith(greek_letter):
-                index = param[len(greek_letter) + 1:]  # Capture the index after _
-                return f'${symbol}_{{{index}}}$'
-        return param.replace("_", " ")
-
-    def create_comparative_plot(self):
-        """Create a comparative plot from selected FRF inputs with custom legends and title"""
-        # Make sure frf_plots is initialized
-        if not hasattr(self, 'frf_plots'):
-            self.frf_plots = {}
-            
-        # Check if we have any FRF data
-        if not self.available_plots_list or self.available_plots_list.count() == 0 or not self.frf_plots:
-            QMessageBox.warning(self, "No Data", "Please run FRF analysis first to generate plots for comparison.")
-            return
-            
-        # Get selected plots
-        selected_items = self.available_plots_list.selectedItems()
-        selected_plot_names = [item.text() for item in selected_items]
-        
-        if not selected_plot_names:
-            QMessageBox.warning(self, "Selection Error", "Please select at least one plot to compare.")
-            return
-            
-        # Check if any of the selected plots exist in frf_plots
-        valid_plot_names = [name for name in selected_plot_names if name in self.frf_plots]
-        if not valid_plot_names:
-            QMessageBox.warning(self, "Data Error", "None of the selected plots have valid FRF data.")
-            return
-        
-        # Update legend map from table
-        self.legend_map = {}
-        for row in range(self.legend_table.rowCount()):
-            original_name = self.legend_table.item(row, 0).text()
-            custom_legend = self.legend_table.item(row, 1).text() if self.legend_table.item(row, 1) else ""
-            
-            # Get style options if available
-            line_style = "-"
-            marker = ""
-            color = None
-            
-            # Get style widgets if they exist
-            line_style_widget = self.legend_table.cellWidget(row, 2)
-            marker_widget = self.legend_table.cellWidget(row, 3)
-            color_widget = self.legend_table.cellWidget(row, 4)
-            
-            if line_style_widget:
-                line_style = line_style_widget.currentText()
-                if line_style == "None":
-                    line_style = ""
-                    
-            if marker_widget:
-                marker = marker_widget.currentText()
-                if marker == "None":
-                    marker = ""
-                    
-            if color_widget:
-                # Get color from the button's background color
-                # The QPushButton doesn't have currentText() method
-                style = color_widget.styleSheet()
-                # Extract color from background-color: X; in the style sheet
-                import re
-                color_match = re.search(r'background-color:\s*(.*?);', style)
-                if color_match:
-                    color = color_match.group(1)
-                else:
-                    # Use default color if can't extract from style
-                    import random
-                    r, g, b = [random.randint(0, 255) for _ in range(3)]
-                    color = f"rgb({r},{g},{b})"
-            
-            # Store in legend map
-            self.legend_map[original_name] = {
-                'legend': custom_legend,
-                'line_style': line_style,
-                'marker': marker,
-                'color': color
-            }
-        
-        # Get custom title
-        custom_title = self.plot_title_edit.text()
-        if not custom_title:
-            custom_title = "Comparative FRF Visualization"
-        
-        # Check if figure exists, create it if not
-        if not hasattr(self, 'comp_fig') or self.comp_fig is None:
-            # Create matplotlib figure
-            from matplotlib.figure import Figure
-            from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
-            
-            # Create the figure
-            self.comp_fig = Figure(figsize=(self.fig_width_spin.value(), self.fig_height_spin.value()))
-            self.comp_canvas = FigureCanvas(self.comp_fig)
-            
-            # Check if we have a layout to add it to
-            if hasattr(self, 'comp_plot_layout') and self.comp_plot_layout is not None:
-                # Remove any existing widgets
-                while self.comp_plot_layout.count():
-                    item = self.comp_plot_layout.takeAt(0)
-                    if item.widget():
-                        item.widget().deleteLater()
-                
-                # Add the new canvas
-                self.comp_plot_layout.addWidget(self.comp_canvas)
-        
-        # Clear previous figure and set its size
-        self.comp_fig.clear()
-        self.comp_fig.set_size_inches(self.fig_width_spin.value(), self.fig_height_spin.value())
-        ax = self.comp_fig.add_subplot(111)
-        
-        # Get normalization values - set to meaningful defaults and print for debugging
-        x_norm_factor = self.x_norm_value.value() if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() else 1.0
-        y_norm_factor = self.y_norm_value.value() if hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked() else 1.0
-        # Debug the normalization factors
-        print(f"\nDEBUG - Normalization factors: X={x_norm_factor}, Y={y_norm_factor}")
-        print(f"Normalization checkboxes: X={hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked()}, Y={hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked()}\n")
-        
-        # Plot each selected FRF
-        for plot_name in selected_plot_names:
-            if plot_name in self.frf_plots:
-                # Extract data from original plot
-                orig_fig = self.frf_plots[plot_name]
-                orig_ax = orig_fig.axes[0]
-                
-                # Extract lines data
-                for line in orig_ax.get_lines():
-                    # Get the data
-                    x_data = line.get_xdata()
-                    y_data = line.get_ydata()
-                    source_line_label = line.get_label() # Original label from the matplotlib line object
-                    # Debug the original data
-                    if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() or hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked():
-                        self.debug_array(x_data, "x_data BEFORE")
-                        self.debug_array(y_data, "y_data BEFORE")
-
-                    # Ensure we're working with numpy arrays
-                    x_data_np = np.array(x_data)
-                    y_data_np = np.array(y_data)
-
-                    # Apply normalization explicitly with numpy
-                    if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() and x_norm_factor != 1.0:
-                        x_data_np = x_data_np / x_norm_factor
-                        print(f"Applied X normalization: divided by {x_norm_factor}")
-
-                    if hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked() and y_norm_factor != 1.0:
-                        y_data_np = y_data_np / y_norm_factor
-                        print(f"Applied Y normalization: divided by {y_norm_factor}")
-
-                    # Always use the numpy arrays for plotting
-                    x_data = x_data_np
-                    y_data = y_data_np
-
-                    # Debug the normalized data
-                    if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() or hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked():
-                        self.debug_array(x_data, "x_data AFTER")
-                        self.debug_array(y_data, "y_data AFTER")
-
-                    # Start with defaults from the original line's properties
-                    current_legend_text = source_line_label
-                    current_line_style = line.get_linestyle()
-                    current_marker = line.get_marker()
-                    current_color = line.get_color() # Matplotlib's auto-assigned color from its cycle
-
-                    # Determine which properties from self.legend_map to use.
-                    # self.legend_map is keyed by names from column 0 of the legend_table,
-                    # which can be either a plot_name (e.g., "Mass 1 (With DVA) - RunX")
-                    # or a source_line_label (e.g., "mass_1").
-                    # We prioritize the specific source_line_label's settings if they exist.
-                    
-                    props_to_use = None
-                    if source_line_label and source_line_label in self.legend_map:
-                        # A specific entry exists for this line's original label
-                        props_to_use = self.legend_map[source_line_label]
-                    elif plot_name in self.legend_map:
-                        # No specific entry for the line's label, fall back to the overall plot_name's entry
-                        props_to_use = self.legend_map[plot_name]
-
-                    if props_to_use:
-                        # Apply custom legend text if 'legend' key explicitly exists in the properties.
-                        # This ensures that an empty string provided by the user is respected.
-                        if 'legend' in props_to_use:
-                            current_legend_text = props_to_use['legend']
-                        
-                        # Apply custom styles, falling back to current (original line) style if not specified in map.
-                        current_line_style = props_to_use.get('line_style', current_line_style)
-                        current_marker = props_to_use.get('marker', current_marker)
-                        
-                        # Get color and convert to matplotlib format if needed
-                        color_value = props_to_use.get('color', current_color)
-                        if isinstance(color_value, str) and color_value.startswith('rgb('):
-                            # Convert 'rgb(r,g,b)' format to a tuple
-                            try:
-                                import re
-                                r, g, b = map(int, re.search(r'rgb\((\d+),(\d+),(\d+)\)', color_value).groups())
-                                current_color = (r/255, g/255, b/255)  # Normalize to 0-1 range
-                            except (ValueError, AttributeError):
-                                # If conversion fails, use the original color
-                                pass
-                        else:
-                            current_color = color_value
-
-                    # Convert "None" string values from comboboxes to what Matplotlib expects
-                    if current_line_style == "None": current_line_style = "" # Empty string for no line
-                    if current_marker == "None": current_marker = "none" # "none" for no marker
-                    
-                    # Plot the line on the comparative figure with customized style
-                    ax.plot(x_data, y_data, label=current_legend_text,
-                            linestyle=current_line_style,
-                            marker=current_marker,
-                            color=current_color) # Matplotlib handles color strings and its default cycle
-        
-        # Set title with custom font size
-        ax.set_title(custom_title, fontsize=self.title_font_size.value())
-        
-        # Set axis labels with normalization factor information
-        x_label = 'Frequency (rad/s)'
-        if hasattr(self, 'x_norm_check') and self.x_norm_check.isChecked() and x_norm_factor != 1.0:
-            x_label += f' / {x_norm_factor}'
-        ax.set_xlabel(x_label)
-
-        y_label = 'Amplitude'
-        if hasattr(self, 'y_norm_check') and self.y_norm_check.isChecked() and y_norm_factor != 1.0:
-            y_label += f' / {y_norm_factor}'
-        ax.set_ylabel(y_label)
-        
-        # Add grid based on checkbox state
-        ax.grid(self.show_grid_check.isChecked(), linestyle='--', linewidth=0.5, alpha=0.7)
-        
-        # Add legend with custom position
-        legend_position = self.legend_position_combo.currentText()
-        ax.legend(loc=legend_position)
-        
-        # Update the figure
-        self.comp_fig.tight_layout()
-        self.comp_canvas.draw()
-        
-        # Update the legend table
-        self._update_legend_table_from_selection()
-        
-        # Show success message
-        self.status_bar.showMessage("Comparative plot created successfully")
-
-
-    def create_catalogue_creator_page(self):
-        """Empty placeholder for removed catalogue creator page"""
-        # Create empty page
-        empty_page = QWidget()
-        self.content_stack.addWidget(empty_page)
